SimilaritySearch.jl

SimilaritySearch.jl is a library for nearest neighbor search. In particular, it contains the implementation for SearchGraph:

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

@article{tellezscalable,
  title={A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs},
  author={Tellez, Eric S and Ruiz, Guillermo and Chavez, Edgar and Graff, Mario},
  journal={Pattern Analysis and Applications},
  pages={1--15},
  publisher={Springer}
}

Installing SimilaritySearch

You may install the package as follows

] add SimilaritySearch

also, you can run the set of tests as fol

] test SimilaritySearch

Using the library

Please see examples directory of this repository. Here you will find a list of Pluto's notebooks that exemplifies its usage.

API

SimilaritySearch.ExhaustiveSearch

SimilaritySearch.PivotedSearch

SimilaritySearch.PivotedSearch

SimilaritySearch.SearchGraph

SimilaritySearch.SearchGraphOptions

Base.empty!

Base.first

Base.last

Base.length

Base.popfirst!

Base.push!

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

Distances.evaluate

SimilaritySearch.LcsDistance

SimilaritySearch.LevenshteinDistance

SimilaritySearch.common_prefix

SimilaritySearch.compute_distances

SimilaritySearch.compute_distances

SimilaritySearch.covrad

SimilaritySearch.distant_tournament

SimilaritySearch.find_neighborhood

SimilaritySearch.find_neighborhood

SimilaritySearch.fix_order!

SimilaritySearch.hill_climbing

SimilaritySearch.k_near_and_far

SimilaritySearch.maxlength

SimilaritySearch.opt_expand_neighborhood

SimilaritySearch.optimize!

SimilaritySearch.optimize!

SimilaritySearch.pop!

SimilaritySearch.push_neighborhood!

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.search

SimilaritySearch.select_sss

SimilaritySearch.select_tournament

SimilaritySearch.sss

SimilaritySearch.union_intersection

SimilaritySearch.ExhaustiveSearch — Type

ExhaustiveSearch(dist::PreMetric, db::AbstractVector, knn::KnnResult)
ExhaustiveSearch(dist::PreMetric, db::AbstractVector, k::Integer)

Defines an exhaustive search

SimilaritySearch.PivotedSearch — Type

PivotedSeach(index::PivotTable, db::AbstractVector, dist::PreMetric, knn::KnnResult)
PivotedSeach(index::PivotTable, db::AbstractVector, dist::PreMetric, k::Integer=10)

Defines a search index for Pivot tables

SimilaritySearch.PivotedSearch — Method

PivotedSearch(::Type{PivotTable}, dist::PreMetric, db::AbstractVector{T}, pivots::Vector{T})
PivotedSearch(::Type{PivotTable}, dist::PreMetric, db::AbstractVector{T}, numpivots::Integer)

Creates a PivotTable index with the given pivots. If the number of pivots is specified, then they will be randomly selected from the dataset.

SimilaritySearch.SearchGraph — Method

SearchGraph(dist::PreMetric, db::AbstractVector;
    search_algo::LocalSearchAlgorithm=BeamSearch(),
    neighborhood_algo::NeighborhoodAlgorithm=LogNeighborhood(),
    automatic_optimization=false,
    recall=0.9,
    ksearch=10,
    tol=0.001,
    verbose=true)

Creates a SearchGraph object, i.e., an index to perform approximate search on db using the given search and neighbohood strategies. If automatic_optimization is true, then the structure tries to reach the given recall under the given ksearch.

SimilaritySearch.SearchGraphOptions — Type

SearchGraphOptions

Defines a number of options for the SearchGraph

Base.empty! — Function

empty!(res::KnnResult)
empty!(res::KnnResult, k::Integer)

Clears the content of the result pool. If k is given then the size of the pool is changed; the internal buffer is adjusted as needed (only grows).

Base.first — Method

first(p::KnnResult)

Return the first item of the result set, the closest item

Base.last — Method

last(p::KnnResult)

Returns the last item of the result set

Base.length — Method

length(p::KnnResult)

length returns the number of items in the result set

Base.popfirst! — Method

popfirst!(p::KnnResult)

Removes and returns the nearest neeighboor from the pool, an O(length(p.pool)) operation

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors.

Distances.evaluate — Method

evaluate(::CosineDistance, a, b)

Computes the cosine distance between two vectors. Please use AngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::DiceDistance, a, b)

Computes the Dice's distance of a and b both sets specified as sorted vectors.

Distances.evaluate — Method

evaluate(GenericLevenshteinDistance, a, b)::Int

Computes the edit distance between two strings, this is a low level function

Distances.evaluate — Method

evaluate(::IntersectionDissimilarity, a, b)

(a, b)

Uses the intersection as a distance function (non-metric)

Distances.evaluate — Method

evaluate(L1Distance, a, b)

Computes the Manhattan's distance between a and b

Distances.evaluate — Method

evaluate(L2Distance, a, b)

Computes the Euclidean's distance betweem a and b

Distances.evaluate — Method

evaluate(::LInftyDistance, a, b)

Computes the max or Chebyshev'se distance

Distances.evaluate — Method

evaluate(lp::LpDistance, a, b)

Computes generic Minkowski's distance

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors. It supposes that all vectors are normalized (see normalize! function)

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a, b)

Computes the cosine distance between two vectors, it expects normalized vectors (see normalize! method). Please use NormalizedAngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::SqL2Distance, a, b)

Computes the squared Euclidean's distance between a and b

Distances.evaluate — Method

evaluate(::BinaryHammingDistance, a, b)::Float64
evaluate(::BinaryHammingDistance, a::AbstractVector, b::AbstractVector)::Float64 where T<:Unsigned

Computes the binary hamming distance for bit types and arrays of bit types

SimilaritySearch.LcsDistance — Method

LcsDistance(a, b)

Instantiates a GenericLevenshteinDistance object to perform LCS distance

SimilaritySearch.LevenshteinDistance — Method

LevenshteinDistance(a, b)

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.common_prefix — Method

common_prefix(a, b)

Computes the length of the common prefix among two strings represented as arrays

SimilaritySearch.compute_distances — Method

compute_distances(dist::PreMetric, db::AbstractVector{T}, refs::AbstractVector{Int}, q::T) where T

Computes the distances of q to the set of references refs (each index point to an item in db) It returns an array of tuples (distance, refID)

SimilaritySearch.compute_distances — Method

compute_distances(dist::PreMetric, refs::AbstractVector{T}, q::T) where T

Computes the distances of q to the set of references refs It returns an array of tuples (distance, refID)

SimilaritySearch.covrad — Method

covrad(p::KnnResult)

Returns the coverage radius of the result set; if length(p) < K then typemax(Float32) is returned

SimilaritySearch.distant_tournament — Method

distant_tournament(dist::PreMetric, db::Array{T,1}, numrefs::Integer, tournamentsize::Integer=3) where T

Creates a pivot table where each pivot is relativaly distant each other based on a distant tournament

SimilaritySearch.find_neighborhood — Method

find_neighborhood(algo::FixedNeighborhood, index::SearchGraph, item)

Finds a list of neighbors using the FixedNeighborhood criterion of item in the index

SimilaritySearch.find_neighborhood — Method

find_neighborhood(index::SearchGraph{T}, item)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.fix_order! — Method

fix_order!(res::KnnResult, n)

Fixes the sorted state of the array. It implements a kind of insertion sort It is efficient due to the expected distribution of the items being inserted (it is expected just a few elements smaller than the current ones)

SimilaritySearch.hill_climbing — Method

hill_climbing(isearch::IHCSearch, index::SearchGraph, q, res::KnnResult, nodeID::Integer)

Runs a single hill climbing search process starting in vertex nodeID

SimilaritySearch.k_near_and_far — Method

k_near_and_far(dist::PreMetric, near::KnnResult, far::KnnResult, obj::T, refs::Vector{T}, k::Integer) where T

Searches for k near and far objects in the set of references

SimilaritySearch.maxlength — Method

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

SimilaritySearch.opt_expand_neighborhood — Method

opt_expand_neighborhood(fun, ihc::IHCSearch, n::Integer, iter::Integer, probes::Integer)

Generates configurations of the IHCSearch that feed the optimize! function (internal function)

SimilaritySearch.optimize! — Method

function optimize!(search_algo::LocalSearchAlgorithm,
                   index::SearchGraph{T},
                   recall::Float64,
                   perf::Performance;
                   bsize::Int=4,
                   tol::Float64=0.01,
                   probes::Int=0) where T

Optimizes a local search index for an specific algorithm to get the desired performance. Note that optimizing for low-recall will yield to faster searches; the train queries are specified as part of the perf struct.

SimilaritySearch.optimize! — Method

optimize!(perf::Performance, index::SearchGraph;
recall=0.9, ksearch=10, verbose=index.opts.verbose, tol::Float64=0.01, maxiters::Integer=3, probes::Integer=0) 
optimize!(perf, index.search_algo, index; recall=recall, tol=tol, maxiters=3, probes=probes)

Optimizes the index for the specified kind of queries.

SimilaritySearch.pop! — Method

pop!(p)

Removes and returns the last item in the pool, it is an O(1) operation

SimilaritySearch.push_neighborhood! — Method

push_neighborhood!(index::SearchGraph, item, L::AbstractVector{Int32})

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

SimilaritySearch.search — Function

search(searchctx::AbstractSearchContext, q, k::Integer=maxlength(searchctx.res))
search(searchctx::AbstractSearchContext, q)

This is the most generic search function. It calls almost all implementations whenever an integer k is given.

SimilaritySearch.search — Method

Tries to reach the set of nearest neighbors specified in res for q.

bs: the parameters of BeamSearch

index: the local search index

q: the query

res: The result object, it stores the results and also specifies the kind of query

SimilaritySearch.search — Method

search(seq::ExhaustiveSearch, q, res::KnnResult)

Solves the query evaluating all items in the given query.

By default, it uses an internal result buffer; multithreading applications must duplicate specify another res object.

SimilaritySearch.search — Method

search(isearch::IHCSearch, index::SearchGraph, q, res::KnnResult)

Performs an iterated hill climbing search for q.

SimilaritySearch.search — Method

search(index::PivotedSeach, q, res::KnnResult)

Solves a query with the pivot index.

SimilaritySearch.search — Method

search(index::SearchGraph, q, res::KnnResult)

Solves the specified query res for the query object q.

SimilaritySearch.search — Method

search(kvp::Kvp, q::T, res::KnnResult) where T

Searches for q in the kvp index

SimilaritySearch.select_sss — Method

select_sss(dist::PreMetric, db::AbstractVector{T}, alpha::Float64; shuf::Bool=true) where T

selectsss selects the necessary pivots to fulfill the SSS criterion using :param:alpha. If :param:shuffledb is true then the database is shuffled before the selection process; in any case, the estimation of the maximum distance introduces indeterminism, however it could be too small. If you need better a better random selection set :param:shuf as true

It returns a set of pivots as a list of integers pointing to elements in :param:db

SimilaritySearch.select_tournament — Method

select_tournament(dist::PreMetric, db::AbstractVector{T}, numrefs::Int, tournamentsize::Int) where T

select_tournament selects numrefs references from db using a tournament criterion; each individual is selected amongtournamentsize individuals.

It returns a set of pivots as a list of integers pointing to elements in db

SimilaritySearch.sss — Method

sss(dist::PreMetric, db::Array{T,1}; alpha::Real=0.35, shuf=false) where T

Creates an SSS pivot table.

SimilaritySearch.union_intersection — Method

union_intersection(a::T, b::T)

Computes both the size of the unions an the size the intersections of a and b; specified as ordered sequences.