Cosine distance functions

SimilaritySearch implements some cosine/angle distance functions. Please recall that we use the evaluate function definition from Distances.jl.

SimilaritySearch.CosineDistance — Type

CosineDistance()

The cosine is defined as:

\[\cos(u, v) = \frac{\sum_i u_i v_i}{\sqrt{\sum_i u_i^2} \sqrt{\sum_i v_i^2}}\]

The cosine distance is defined as $1 - \cos(u,v)$

SimilaritySearch.AngleDistance — Type

AngleDistance()

The angle distance is defined as:

\[∠(u, v)= \arccos(\cos(u, v))\]

SimilaritySearch.NormalizedCosineDistance — Type

NormalizedCosineDistance()

Similar to CosineDistance but suppose that input vectors are already normalized

\[1 - \sum_i {u_i v_i}\]

SimilaritySearch.NormalizedAngleDistance — Type

NormalizedAngleDistance()

Similar to AngleDistance but suppose that input vectors are already normalized

\[\arccos \sum_i {u_i v_i}\]