Hamming distance functions

The hamming distance for binary and string data is implemented. Please recall that we use the evaluate function definition from Distances.jl.

SimilaritySearch.BinaryHammingDistance — Type

BinaryHammingDistance()

Binary hamming uses bit wise operations to count the differences between bit strings

SimilaritySearch.StringHammingDistance — Type

StringHammingDistance()

The hamming distance counts the differences between two equally sized strings