Set distances

The following set distances are supported. Please recall that we use the evaluate function definition from Distances.jl.

SimilaritySearch.JaccardDistance — Type

JaccardDistance()

The Jaccard distance is defined as

\[J(u, v) = \frac{|u \cap v|}{|u \cup v|}\]

SimilaritySearch.DiceDistance — Type

DiceDistance()

The Dice distance is defined as

\[D(u, v) = \frac{2 |u \cap v|}{|u| + |v|}\]

SimilaritySearch.IntersectionDissimilarity — Type

IntersectionDissimilarity()

The intersection dissimilarity uses the size of the intersection as a mesuare of similarity as follows:

\[I(u, v) = 1 - \frac{|u \cap v|}{\max \{|u|, |v|\}}\]