String distances

The following string distances are supported. Please recall that we use the evaluate function definition from Distances.jl.

SimilaritySearch.LevenshteinDistance — Function

LevenshteinDistance(a, b)

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.LcsDistance — Function

LcsDistance(a, b)

Instantiates a GenericLevenshteinDistance object to perform LCS distance

SimilaritySearch.GenericLevenshteinDistance — Type

GenericLevenshteinDistance(icost, dcost, rcost)

The levenshtein distance measures the minimum number of edit operations to convert one string into another. The costs insertion icost, deletion cost dcost, and replace cost rcost.

SimilaritySearch.CommonPrefixDissimilarity — Type

CommonPrefixDissimilarity()

Uses the common prefix as a measure of dissimilarity between two strings