KnnResult

SimilaritySearch's core is to solve knn searches; for this matter, it relies on the KnnResult struct and its related functions.

SimilaritySearch.Item — Type

Item(id, dist)

An item identifier and its related distance to another item

SimilaritySearch.KnnResult — Type

KnnResult(ksearch::Integer)
KnnResult(arrOfItems::AbstractVector)
KnnResult(currsize, capacity, pool)  # low level constructor

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

SimilaritySearch.maxlength — Function

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

SimilaritySearch.covrad — Function

covrad(p::KnnResult)

Returns the coverage radius of the result set; if length(p) < K then typemax(Float32) is returned

Missing docstring.

Missing docstring for maxlength. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(res::KnnResult, id::Integer, dist::AbstractFloat). Check Documenter's build log for details.