Indexes

SimilaritySearch.ExhaustiveSearch — Type

ExhaustiveSearch(dist::SemiMetric, db::AbstractVector)

Solves queries evaluating dist for the query and all elements in the dataset

SimilaritySearch.SearchGraph — Type

struct SearchGraph <: AbstractSearchContext

SearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.

hints: Initial points for exploration (empty hints imply using random points)

Note: Parallel insertions should be made through append! or index! function with parallel_block > 1

Searching

SimilaritySearch.search — Function

search(seq::ExhaustiveSearch, q, res::KnnResult)

Solves the query evaluating all items in the given query.

search(bs::BeamSearch, index::SearchGraph, q, res, hints, pools; bsize=bs.bsize, Δ=bs.Δ, maxvisits=bs.maxvisits)

Tries to reach the set of nearest neighbors specified in res for q.

bs: the parameters of BeamSearch

index: the local search index

q: the query

res: The result object, it stores the results and also specifies the kind of query

hints: Starting points for searching, randomly selected when it is an empty collection

pools: A SearchGraphPools object with preallocated pools

Optional arguments (defaults to values in bs)

bsize: Beam size

Δ: exploration expansion factor

maxvisits: Maximum number of nodes to visit (distance evaluations)

search(index::SearchGraph, q, res; hints=index.hints, pools=getpools(index))

Solves the specified query res for the query object q.

SimilaritySearch.searchbatch — Function

searchbatch(index, Q, k::Integer=10; parallel=false, pools=GlobalKnnResult) -> indices, distances

Searches a batch of queries in the given index (searches for k neighbors).

parallel specifies if the query should be solved in parallel at object level (each query is sequentially solved but many queries solved in different threads).

pool relevant if parallel=true. If it is explicitly given it should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations. In most case uses the default is enought, but different pools should be used when indexes use internal indexes to solve queries (e.g., using index's proxies or database objects defined as indexes).

Note: The i-th column in indices and distances correspond to the i-th query in Q Note: The final indices at each column can be 0 if the search process was unable to retrieve k neighbors.

searchbatch(index, Q, I::AbstractMatrix{Int32}, D::AbstractMatrix{Float32}; parallel=false, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index and I and D as output (searches for k=size(I, 1))

searchbatch(index, Q, KNN::AbstractVector{KnnResult}; parallel=false, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index using an array of KnnResult's; each KnnResult object can specify different k values.

Computing all knns

The operation of computing all knns in the index is computed as follows:

SimilaritySearch.allknn — Function

allknn(g::AbstractSearchContext, k::Integer; parallel=false, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Parameters:

g: the index

k: the number of neighbors to retrieve

parallel: If true, the construction will use all threads available threads

pools: A pools object, dependent of g

Returns:

knns a (k, n) matrix of identifiers; the i-th column corresponds to the i-th object in the dataset. Zeros can happen to the end of each column meaning that the retrieval was less than the desired k

dists a (k, n) matrix of distances; the i-th column corresponds to the i-th object in the dataset. Zero values in knns should be ignored in dists

allknn(g, knns, dists; parallel=false, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Arguments:

g: the index

knns: an uninitialized integer matrix of (k, n) size for storing the k nearest neighbors of the n elements

dists: an uninitialized floating point matrix of (k, n) size for storing the k nearest distances of the n elements

parallel: If true, the construction will use all threads available threads

pools: A pools object, dependent of g

Results:

knns and dists are returned. Note that the index can retrieve less than k objects, and these are represented as zeros at the end of each column (can happen)

Indexing elements

Base.push! — Method

push!(
    index::SearchGraph,
    item;
    push_item=true,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends an object into the index. It accepts the same arguments that push! but assuming some default values.

Arguments:

index: The search graph index where the insertion is going to happen

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

push_item: if push_item=false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed)

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: callbacks=nothing ignores the execution of any callback

SimilaritySearch.push_item! — Function

pushitem!( index::SearchGraph, item, pushitem, callbacks, pools )

Appends an object into the index

Arguments:

index: The search graph index where the insertion is going to happen

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

push_item: if false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed)

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: setting callbacks as nothing ignores the execution of any callback

Base.append! — Function

append!(
    index::SearchGraph,
    db;
    parallel_block=1,
    parallel_minimum_first_block=parallel_block,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends all items in db to the index. It can be made in parallel or sequentially.

Arguments:

index: the search graph index

db: the collection of objects to insert, an AbstractDatabase is the canonical input, but supports any iterable objects

parallel_block: The number of elements that the multithreading algorithm process at once, it is important to be larger that the number of available threads but not so large since the quality of the search graph could degrade (a few times the number of threads is enough). If parallel_block=1 the algorithm becomes sequential.

parallel_minimum_first_block: The number of sequential appends before running parallel.

Note: Parallel doesn't trigger callbacks inside blocks.

callbacks: A SearchGraphCallbacks object to be called after some insertions (specified by the callbacks object). These callbacks are used to maintain the algorithm in good shape after many insertions (adjust hyperparameters and the structure).

pools: The set of caches used for searching.

Note 1: Callbacks are not executed inside parallel blocks Note 2: Callbacks will be ignored if callbacks=nothing

SimilaritySearch.index! — Function

index!(index::SearchGraph; parallel_block=1, parallel_minimum_first_block=parallel_block, callbacks=SearchGraphCallbacks())

Indexes the already initialized database (e.g., given in the constructor method). It can be made in parallel or sequentially. The arguments are the same than append! function but using the internal index.db as input.

Distance functions

The distance functions are defined to work under the evaluate(::metric, u, v) function (borrowed from Distances.jl package).

Cosine and angle distances for vectors

SimilaritySearch.CosineDistance — Type

CosineDistance()

The cosine is defined as:

\[\cos(u, v) = \frac{\sum_i u_i v_i}{\sqrt{\sum_i u_i^2} \sqrt{\sum_i v_i^2}}\]

The cosine distance is defined as $1 - \cos(u,v)$

SimilaritySearch.NormalizedCosineDistance — Type

NormalizedCosineDistance()

Similar to CosineDistance but suppose that input vectors are already normalized, and therefore, reduced to simply one minus the dot product.

\[1 - \sum_i {u_i v_i}\]

SimilaritySearch.AngleDistance — Type

AngleDistance()

The angle distance is defined as:

\[∠(u, v)= \arccos(\cos(u, v))\]

SimilaritySearch.NormalizedAngleDistance — Type

NormalizedAngleDistance()

Similar to AngleDistance but suppose that input vectors are already normalized

\[\arccos \sum_i {u_i v_i}\]

Set distance functions

Set bbject are represented as ordered arrays

SimilaritySearch.JaccardDistance — Type

JaccardDistance()

The Jaccard distance is defined as

\[J(u, v) = \frac{|u \cap v|}{|u \cup v|}\]

SimilaritySearch.DiceDistance — Type

DiceDistance()

The Dice distance is defined as

\[D(u, v) = \frac{2 |u \cap v|}{|u| + |v|}\]

SimilaritySearch.IntersectionDissimilarity — Type

IntersectionDissimilarity()

The intersection dissimilarity uses the size of the intersection as a mesuare of similarity as follows:

\[I(u, v) = 1 - \frac{|u \cap v|}{\max \{|u|, |v|\}}\]

SimilaritySearch.CosineDistanceSet — Type

CosineDistanceSet()

The cosine distance for very sparse binary vectors represented as sorted lists of positive integers where ones occur.

String alignment distances

The following uses strings/arrays as input, i.e., objects follow the array interface. A broader set of distances for strings can be found in the StringDistances.jl package.

SimilaritySearch.CommonPrefixDissimilarity — Type

CommonPrefixDissimilarity()

Uses the common prefix as a measure of dissimilarity between two strings

SimilaritySearch.GenericLevenshteinDistance — Type

GenericLevenshteinDistance(;icost, dcost, rcost)

The levenshtein distance measures the minimum number of edit operations to convert one string into another. The costs insertion icost, deletion cost dcost, and replace cost rcost. Not thread safe, use a copy of for each thread.

SimilaritySearch.StringHammingDistance — Type

StringHammingDistance()

The hamming distance counts the differences between two equally sized strings

SimilaritySearch.LevenshteinDistance — Function

LevenshteinDistance()

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.LcsDistance — Function

LcsDistance()

Instantiates a GenericLevenshteinDistance object to perform LCS distance

Distances for Cloud of points

SimilaritySearch.HausdorffDistance — Type

HausdorffDistance(dist::SemiMetric)

Hausdorff distance is defined as the maximum of the minimum between two clouds of points.

\[Hausdorff(U, V) = \max{\max_{u \in U} nndist(u, V), \max{v \in V} nndist(v, U) }\]

where $nndist(u, V)$ computes the distance of $u$ to its nearest neighbor in $V$ using the dist metric.

SimilaritySearch.MinHausdorffDistance — Type

MinHausdorffDistance(dist::SemiMetric)

Similar to HausdorffDistance but using minimum instead of maximum.

Public API

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors.

Distances.evaluate — Method

evaluate(::BinaryHammingDistance, a, b)
evaluate(::BinaryHammingDistance, a::AbstractVector, b::AbstractVector) where {T<:Unsigned}

Computes the binary hamming distance for bit types and arrays of bit types

Distances.evaluate — Method

evaluate(::CommonPrefixDissimilarity, a, b)

Computes a dissimilarity based on the common prefix between two strings

Distances.evaluate — Method

evaluate(::CosineDistance, a, b)

Computes the cosine distance between two vectors. Please use AngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::CosineDistanceSet, a, b)

Computes the cosine distance where a and b are sorted lists of integers (emulating binary sparse vectores)

Distances.evaluate — Method

evaluate(::DiceDistance, a, b)

Computes the Dice's distance of a and b both sets specified as sorted vectors.

Distances.evaluate — Method

evaluate(::GenericLevenshteinDistance, a, b)

Computes the edit distance between two strings, this is a low level function

Distances.evaluate — Method

evaluate(m::HausdorffDistance, u, v)

Computes the Hausdorff distance between two cloud of points.

u and v are iterables where each object can be measured with the internal distance dist

Distances.evaluate — Method

evaluate(::IntersectionDissimilarity, a, b)

Uses the intersection as a distance function (non-metric)

Distances.evaluate — Method

evaluate(::JaccardDistance, a, b)

Computes the Jaccard's distance of a and b both sets specified as sorted vectors.

Distances.evaluate — Method

evaluate(::L1Distance, a, b)

Computes the Manhattan's distance between a and b

Distances.evaluate — Method

evaluate(::L2Distance, a, b)

Computes the Euclidean's distance betweem a and b

Distances.evaluate — Method

evaluate(::LInftyDistance, a, b)

Computes the maximum distance or Chebyshev's distance

Distances.evaluate — Method

evaluate(lp::LpDistance, a, b)

Computes generic Minkowski's distance

Distances.evaluate — Method

evaluate(m::MinHausdorffDistance, u, v)

Computes a variant of the Hausdorff distance that uses the minimum instead of the maximum. u and v are iterables where each object can be measured with the internal distance dist

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors. It supposes that all vectors are normalized

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a, b)

Computes the cosine distance between two vectors, it expects normalized vectors. Please use NormalizedAngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::SqL2Distance, a, b)

Computes the squared Euclidean's distance between a and b

Distances.evaluate — Method

 evaluate(::StringHammingDistance, a, b)

Computes the hamming distance between two sequences of the same length

SimilaritySearch.LcsDistance — Method

LcsDistance()

Instantiates a GenericLevenshteinDistance object to perform LCS distance

SimilaritySearch.LevenshteinDistance — Method

LevenshteinDistance()

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.allknn — Method

allknn(g, knns, dists; parallel=false, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Arguments:

g: the index

knns: an uninitialized integer matrix of (k, n) size for storing the k nearest neighbors of the n elements

dists: an uninitialized floating point matrix of (k, n) size for storing the k nearest distances of the n elements

parallel: If true, the construction will use all threads available threads

pools: A pools object, dependent of g

Results:

knns and dists are returned. Note that the index can retrieve less than k objects, and these are represented as zeros at the end of each column (can happen)

SimilaritySearch.allknn — Method

allknn(g::AbstractSearchContext, k::Integer; parallel=false, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Parameters:

g: the index

k: the number of neighbors to retrieve

parallel: If true, the construction will use all threads available threads

pools: A pools object, dependent of g

Returns:

knns a (k, n) matrix of identifiers; the i-th column corresponds to the i-th object in the dataset. Zeros can happen to the end of each column meaning that the retrieval was less than the desired k

dists a (k, n) matrix of distances; the i-th column corresponds to the i-th object in the dataset. Zero values in knns should be ignored in dists

SimilaritySearch.find_neighborhood — Method

find_neighborhood(index::SearchGraph{T}, item, pools; hints=index.hints)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function). res is always reused since reduce creates a new KnnResult from it (a copy if reduce in its simpler terms)

SimilaritySearch.getpair — Method

getindex(res::KnnResult, i)

Access the i-th item in res

SimilaritySearch.getpair — Method

getindex(res::KnnResultShift, st::KnnResultState, i)

Access the i-th item in res

SimilaritySearch.index! — Method

index!(index::SearchGraph; parallel_block=1, parallel_minimum_first_block=parallel_block, callbacks=SearchGraphCallbacks())

Indexes the already initialized database (e.g., given in the constructor method). It can be made in parallel or sequentially. The arguments are the same than append! function but using the internal index.db as input.

SimilaritySearch.macrorecall — Function

macrorecall(goldI::AbstractMatrix, resI::AbstractMatrix, k=size(goldI, 1))::Float64

Computes the macro recall score using goldI as gold standard and resI as predictions; it expects that matrices of integers (identifiers). If k is given, then the results are cut to first k.

SimilaritySearch.macrorecall — Method

macrorecall(goldlist::AbstractVector, reslist::AbstractVector)::Float64

Computes the macro recall score using sets of results (KnnResult objects or vectors of itegers).

SimilaritySearch.maxlength — Method

maxlength(res::KnnResultShift)

The maximum allowed cardinality (the k of knn)

SimilaritySearch.maxlength — Method

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

SimilaritySearch.optimize! — Method

optimize!(
    index::SearchGraph,
    opt::OptimizeParameters;
    queries=nothing,
    scalemaxvisits=1.0,
    maxvisits=2 * index.search_algo.maxvisits,
    verbose=index.verbose
)

Optimizes the index using the opt parameters. If queries=nothing then it selects a small sample randomply from the already indexed objects; this sample size and the number of neighbors are also parameters in opt.

Note: The visits parameters is a factor allowing early stopping based on the number of distance evaluations seen in the optimization procedure; please note that in some cases this will reduce the quality to limit the search times. Please also take into account that inserting items after limiting visits may also cause severe quality degradation when maxvisits is not also updated as required. You can always adjust maxvisits modifying index.search_algo.maxvisits.

SimilaritySearch.push_item! — Method

pushitem!( index::SearchGraph, item, pushitem, callbacks, pools )

Appends an object into the index

Arguments:

index: The search graph index where the insertion is going to happen

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

push_item: if false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed)

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: setting callbacks as nothing ignores the execution of any callback

SimilaritySearch.push_neighborhood! — Method

push_neighborhood!(index::SearchGraph, item, neighbors, callbacks; push_item=true)

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

SimilaritySearch.recallscore — Method

recallscore(gold::Set, res)

Compute recall and precision scores from the result sets.

SimilaritySearch.reuse! — Function

reuse!(res::KnnResultShift)
reuse!(res::KnnResultShift, k::Integer)

Returns a result set and a new initial state; reuse the memory buffers

SimilaritySearch.reuse! — Function

reuse!(res::KnnResult)
reuse!(res::KnnResult, k::Integer)

Returns a result set and a new initial state; reuse the memory buffers

SimilaritySearch.search — Method

search(bs::BeamSearch, index::SearchGraph, q, res, hints, pools; bsize=bs.bsize, Δ=bs.Δ, maxvisits=bs.maxvisits)

Tries to reach the set of nearest neighbors specified in res for q.

bs: the parameters of BeamSearch

index: the local search index

q: the query

res: The result object, it stores the results and also specifies the kind of query

hints: Starting points for searching, randomly selected when it is an empty collection

pools: A SearchGraphPools object with preallocated pools

Optional arguments (defaults to values in bs)

bsize: Beam size

Δ: exploration expansion factor

maxvisits: Maximum number of nodes to visit (distance evaluations)

SimilaritySearch.search — Method

search(seq::ExhaustiveSearch, q, res::KnnResult)

Solves the query evaluating all items in the given query.

SimilaritySearch.search — Method

search(index::SearchGraph, q, res; hints=index.hints, pools=getpools(index))

Solves the specified query res for the query object q.

SimilaritySearch.searchbatch — Function

searchbatch(index, Q, k::Integer=10; parallel=false, pools=GlobalKnnResult) -> indices, distances

Searches a batch of queries in the given index (searches for k neighbors).

parallel specifies if the query should be solved in parallel at object level (each query is sequentially solved but many queries solved in different threads).

pool relevant if parallel=true. If it is explicitly given it should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations. In most case uses the default is enought, but different pools should be used when indexes use internal indexes to solve queries (e.g., using index's proxies or database objects defined as indexes).

Note: The i-th column in indices and distances correspond to the i-th query in Q Note: The final indices at each column can be 0 if the search process was unable to retrieve k neighbors.

SimilaritySearch.searchbatch — Method

searchbatch(index, Q, I::AbstractMatrix{Int32}, D::AbstractMatrix{Float32}; parallel=false, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index and I and D as output (searches for k=size(I, 1))

SimilaritySearch.searchbatch — Method

searchbatch(index, Q, KNN::AbstractVector{KnnResult}; parallel=false, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index using an array of KnnResult's; each KnnResult object can specify different k values.

SimilaritySearch.timedsearchbatch — Method

timedsearchbatch(index, Q, ksearch::Integer; parallel=false)

Computes the K nearest neigbors of each object in Q and returns two matrices, and the average search time in seconds.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, Δ::Float32)

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time.

bsize: The size of the beam.

Δ: Soft margin for accepting elements into the beam

maxvisits: MAximum visits while searching, useful for early stopping without convergence

SimilaritySearch.BinaryHammingDistance — Type

BinaryHammingDistance()

Binary hamming uses bit wise operations to count the differences between bit strings

SimilaritySearch.Callback — Type

abstract type Callback end

Abstract type to trigger callbacks after some number of insertions. SearchGraph stores the callbacks in callbacks (a dictionary that associates symbols and callback objects); A SearchGraph object controls when callbacks are fired using callback_logbase and callback_starting

SimilaritySearch.DisjointHints — Type

mutable struct DisjointHints

Indicates that hints are a small disjoint (untouched neighbors) subsample

SimilaritySearch.DistalSatNeighborhood — Type

DistalSatNeighborhood()

New items are connected with a small set of items computed with a Distal SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage but in reverse order of distance.

SimilaritySearch.IdentityNeighborhood — Type

struct IdentityNeighborhood

It does not modifies the given neighborhood

SimilaritySearch.KDisjointHints — Type

struct KDisjointHints

Indicates that hints are selected to have a disjoint neighborhood

SimilaritySearch.KnnResult — Type

KnnResult(ksearch::Integer)

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

SimilaritySearch.KnnResultShift — Type

KnnResultShift(ksearch::Integer)

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

SimilaritySearch.L1Distance — Type

L1Distance()

The manhattan distance or $L_1$ is defined as

\[L_1(u, v) = \sum_i{|u_i - v_i|}\]

SimilaritySearch.L2Distance — Type

L2Distance()

The euclidean distance or $L_2$ is defined as

\[L_2(u, v) = \sqrt{\sum_i{(u_i - v_i)^2}}\]

SimilaritySearch.LInftyDistance — Type

LInftyDistance()

The Chebyshev or $L_{\infty}$ distance is defined as

\[L_{\infty}(u, v) = \max_i{\left| u_i - v_i \right|}\]

SimilaritySearch.LpDistance — Type

LpDistance(p)
LpDistance(p, pinv)

The general Minkowski distance $L_p$ distance is defined as

\[L_p(u, v) = \left|\sum_i{(u_i - v_i)^p}\right|^{1/p}\]

Where $p_{inv} = 1/p$. Note that you can specify unrelated p and pinv if you need an specific behaviour.

SimilaritySearch.NeighborhoodReduction — Type

abstract type NeighborhoodReduction end

Overrides Base.reduce(::NeighborhoodReduction, res::KnnResult, index::SearchGraph) to postprocess res using some criteria. Called from find_neighborhood, and returns a new KnnResult struct (perhaps a copy of res) since push_neighborhood captures the reference of its output.

SimilaritySearch.RandomHints — Type

mutable struct RandomHints

Indicates that hints are a random sample of the dataset

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood()

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

SimilaritySearch.SearchGraphPools — Type

SearchGraphPools(results=GlobalKnnResult, vstates=GlobalVisitedVertices, beams=GlobalBeamKnnResult)

A set of pools to alleviate memory allocations in SearchGraph construction and searching. Relevant on multithreading scenarious where distance functions, evaluate can call other metric indexes that can use these shared resources (globally defined).

Each pool is a vector of Threads.nthreads() preallocated objects of the required type.

SimilaritySearch.SqL2Distance — Type

SqL2Distance()

The squared euclidean distance is defined as

\[L_2(u, v) = \sum_i{(u_i - v_i)^2}\]

It avoids the computation of the square root and should be used whenever you are able do it.

Private API

Base.append! — Method

append!(
    index::SearchGraph,
    db;
    parallel_block=1,
    parallel_minimum_first_block=parallel_block,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends all items in db to the index. It can be made in parallel or sequentially.

Arguments:

index: the search graph index

db: the collection of objects to insert, an AbstractDatabase is the canonical input, but supports any iterable objects

parallel_block: The number of elements that the multithreading algorithm process at once, it is important to be larger that the number of available threads but not so large since the quality of the search graph could degrade (a few times the number of threads is enough). If parallel_block=1 the algorithm becomes sequential.

parallel_minimum_first_block: The number of sequential appends before running parallel.

Note: Parallel doesn't trigger callbacks inside blocks.

callbacks: A SearchGraphCallbacks object to be called after some insertions (specified by the callbacks object). These callbacks are used to maintain the algorithm in good shape after many insertions (adjust hyperparameters and the structure).

pools: The set of caches used for searching.

Note 1: Callbacks are not executed inside parallel blocks Note 2: Callbacks will be ignored if callbacks=nothing

Base.iterate — Function

Base.iterate(res::KnnResult, state::Int=1)

Support for iteration

Base.pop! — Method

pop!(res::KnnResultShift, st::KnnResultState)

Removes and returns the last item in the pool, it is an O(1) operation

Base.pop! — Method

pop!(res::KnnResult)

Removes and returns the last item in the pool, it is an O(1) operation

Base.popfirst! — Method

popfirst!(p::KnnResultShift, st::KnnResultState)

Removes and returns the nearest neeighboor pair from the pool, an O(length(p.pool)) operation

Base.popfirst! — Method

popfirst!(p::KnnResult)

Removes and returns the nearest neeighboor pair from the pool, an O(length(p.pool)) operation

Base.push! — Method

push!(res::KnnResult, item::Pair)
push!(res::KnnResult, id::Integer, dist::Real)

Appends an item into the result set

Base.push! — Method

push!(res::KnnResultShift, item::Pair)
push!(res::KnnResultShift, id::Integer, dist::Real)

Appends an item into the result set

SimilaritySearch._find_start_position — Method

Base.iterate(res::KnnResultShift, state::Int=1)

Support for iteration

SimilaritySearch._shifted_fixorder! — Function

_shifted_fixorder!(res, shift=0)

Sorts the result in place; the possible element out of order is on the last entry always. It implements a kind of insertion sort that it is efficient due to the expected distribution of the items being inserted (it is expected just a few elements smaller than the current ones)

SimilaritySearch.common_prefix — Method

common_prefix(a, b)

Computes the length of the common prefix among two strings represented as arrays

SimilaritySearch.execute_callback — Method

execute_callback(opt::KDisjointHints, index)

SearchGraph's callback for selecting hints at random

SimilaritySearch.execute_callback — Method

execute_callback(opt::NeighborhoodSize, index)

SearchGraph's callback for adjusting neighborhood strategy

SimilaritySearch.execute_callback — Method

execute_callback(opt::OptimizeParameters, index::SearchGraph)

SearchGraph's callback for adjunting search parameters

SimilaritySearch.execute_callback — Method

executed_callback(opt::RandomHints, index)

SearchGraph's callback for selecting hints at random

SimilaritySearch.execute_callbacks — Function

execute_callbacks(index::SearchGraph, n=length(index), m=n+1)

Process all registered callbacks

SimilaritySearch.getknnresult — Method

getknnresult(k::Integer, pools) -> KnnResult

Internal function to share result sets for the same thread and avoid memory allocations.

SimilaritySearch.intersectionsize — Function

intersectionsize(a, b, o=Forward)

Computes the size the intersections of a and b, specified as ordered sequences.

SimilaritySearch.reduce_neighborhood — Method

reduce(sat::DistalSatNeighborhood, index::SearchGraph, item, res, pools)

Reduces res using the DistSAT strategy.

SimilaritySearch.unionsize — Method

unionsize(a, b, isize)

Computes the size of the union of a and b that have an intersection size isize

SimilaritySearch.Neighborhood — Type

@with_kw mutable struct Neighborhood

Determines the size of the neighborhood, $k$ is adjusted as a callback, and it is intended to affect previously inserted vertices. The neighborhood is designed to consider two components $k=in+out$, i.e. incoming and outgoing edges for each vertex.

The $out$ size is computed as $minsize + \log(logbase, n)$ where $n$ is the current number of indexed elements; this is computed searching

for $out$ elements in the current index.

The $in$ size is unbounded.

reduce is intended to postprocess neighbors (after search process, i.e., once out edges are computed); do not change $k$ but always must return a copy of the reduced result set.

Note: Set $logbase=Inf$ to obtain a fixed number of $in$ nodes; and set $minsize=0$ to obtain a pure logarithmic growing neighborhood.