Indexes

SimilaritySearch.ExhaustiveSearch — Type

ExhaustiveSearch(dist::SemiMetric, db::AbstractVector)

Solves queries evaluating dist for the query and all elements in the dataset

SimilaritySearch.ParallelExhaustiveSearch — Type

ParallelExhaustiveSearch(; dist=SqL2Distance(), db=VectorDatabase{Float32}())

Solves queries evaluating dist in parallel for the query and all elements in the dataset. Note that this should not be used in conjunction with searchbatch(...; parallel=true) since they will compete for resources.

SimilaritySearch.SearchGraph — Type

struct SearchGraph <: AbstractSearchContext

SearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.

hints: Initial points for exploration (empty hints imply using random points)

Note: Parallel insertions should be made through append! or index! function with parallel_block > 1

Searching

SimilaritySearch.search — Function

search(seq::ExhaustiveSearch, q, res::KnnResult)

Solves the query evaluating all items in the given query.

search(ex::ParallelExhaustiveSearch, q, res::KnnResult; minbatch=0, pools=nothing)

Solves the query evaluating all items in the given query.

Arguments

ex: the search structure

q: the query to solve

res: the result set

Keyword arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: The set of caches (nothing for this index)

search(bs::BeamSearch, index::SearchGraph, q, res, hints, pools; bsize=bs.bsize, Δ=bs.Δ, maxvisits=bs.maxvisits)

Tries to reach the set of nearest neighbors specified in res for q.

bs: the parameters of BeamSearch

index: the local search index

q: the query

res: The result object, it stores the results and also specifies the kind of query

hints: Starting points for searching, randomly selected when it is an empty collection

pools: A SearchGraphPools object with preallocated pools

Optional arguments (defaults to values in bs)

bsize: Beam size

Δ: exploration expansion factor

maxvisits: Maximum number of nodes to visit (distance evaluations)

search(index::SearchGraph, q, res; hints=index.hints, pools=getpools(index))

Solves the specified query res for the query object q.

SimilaritySearch.searchbatch — Function

searchbatch(index, Q, k::Integer; minbatch=0, pools=GlobalKnnResult) -> indices, distances

Searches a batch of queries in the given index (searches for k neighbors).

Arguments

index: The search structure

Q: The set of queries

k: The number of neighbors to retrieve

Keyword arguments

minbatch specifies how many queries are solved per thread.

Integers $1 ≤ minbatch ≤ |Q|$ are valid values

Set minbatch=0 to compute a default number based on the number of available cores.

Set minbatch=-1 to avoid parallelism.

pools relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

Note: The i-th column in indices and distances correspond to the i-th query in Q Note: The final indices at each column can be 0 if the search process was unable to retrieve k neighbors.

searchbatch(index, Q, I::AbstractMatrix{Int32}, D::AbstractMatrix{Float32}; minbatch=0, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index and I and D as output (searches for k=size(I, 1))

Arguments

index: The search structure

Q: The set of queries

k: The number of neighbors to retrieve

Keyword arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

searchbatch(index, Q, KNN::AbstractVector{KnnResult}; minbatch=0, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index using an array of KnnResult's; each KnnResult object can specify different k values.

Arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

Note: KnnResult based functions are significantly faster in general on pre-allocated objects that similar functions accepting matrices of identifiers and distances. Matrix based outputs are based on KnnResult methods that copy their results on the matrices. Preallocation is also costly, so if you have relatively small datasets, you are not intended to repeat the search process many times, or you are unsure, it is safe to use matrix-based functions.

Computing all knns

The operation of computing all knns in the index is computed as follows:

SimilaritySearch.allknn — Function

allknn(g::AbstractSearchContext, k::Integer; minbatch=0, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Parameters:

g: the index

k: the number of neighbors to retrieve

minbatch: controls how multithreading is used for evaluating configurations, see getminbatch

pools: A pools object, dependent of g

Returns:

knns a (k, n) matrix of identifiers; the i-th column corresponds to the i-th object in the dataset. Zeros can happen to the end of each column meaning that the retrieval was less than the desired k

dists a (k, n) matrix of distances; the i-th column corresponds to the i-th object in the dataset. Zero values in knns should be ignored in dists

allknn(g, knns, dists; parallel_block=512, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Arguments:

g: the index

knns: an uninitialized integer matrix of (k, n) size for storing the k nearest neighbors of the n elements

dists: an uninitialized floating point matrix of (k, n) size for storing the k nearest distances of the n elements

minbatch: controls how multithreading is used for evaluating configurations, see getminbatch

pools: A pools object, dependent of g

Results:

knns and dists are returned. Note that the index can retrieve less than k objects, and these are represented as zeros at the end of each column (can happen)

Computing closest pair

The operation of finding the closest pair of elements in the indexed dataset.

SimilaritySearch.closestpair — Function

closestpair(idx::AbstractSearchContext; parallel=false, pools=getpools(idx))

Finds the closest pair among all elements in idx. If the index idx is approximate then pair of points could be also an approximation.

Arguments:

idx: the search structure that indexes the set of points

Keyword Arguments:

parallel: If true then the algorithm uses all available threads to compute the closest pair

pools: The pools needed for the index. Only used for special cases, default values should work in most cases. See getpools for more information.

Indexing elements

Base.push! — Method

push!(
    index::SearchGraph,
    item;
    neighborhood=Neighborhood(),
    push_item=true,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends an object into the index. It accepts the same arguments that push! but assuming some default values.

Arguments:

index: The search graph index where the insertion is going to happen

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

neighborhood: A Neighborhood object that specifies the kind of neighborhood that will be computed.

push_item: if push_item=false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed)

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: callbacks=nothing ignores the execution of any callback

SimilaritySearch.push_item! — Function

push_item!(
    index::SearchGraph,
    item,
    neighborhood,
    push_item,
    callbacks,
    pools
)

Appends an object into the index

Arguments:

index: The search graph index where the insertion is going to happen.

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

neighborhood: A Neighborhood object that specifies the kind of neighborhood that will be computed.

push_item: if false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed).

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: setting callbacks as nothing ignores the execution of any callback

Base.append! — Function

append!(
    index::SearchGraph,
    db;
    neighborhood=Neighborhood(),
    parallel_block=1,
    parallel_minimum_first_block=parallel_block,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends all items in db to the index. It can be made in parallel or sequentially.

Arguments:

index: the search graph index

neighborhood: A Neighborhood object that specifies the kind of neighborhood that will be computed.

db: the collection of objects to insert, an AbstractDatabase is the canonical input, but supports any iterable objects

parallel_block: The number of elements that the multithreading algorithm process at once, it is important to be larger that the number of available threads but not so large since the quality of the search graph could degrade (a few times the number of threads is enough). If parallel_block=1 the algorithm becomes sequential.

parallel_minimum_first_block: The number of sequential appends before running parallel.

Note: Parallel doesn't trigger callbacks inside blocks.

callbacks: A SearchGraphCallbacks object to be called after some insertions (specified by the callbacks object). These callbacks are used to maintain the algorithm in good shape after many insertions (adjust hyperparameters and the structure).

pools: The set of caches used for searching.

Note 1: Callbacks are not executed inside parallel blocks Note 2: Callbacks will be ignored if callbacks=nothing

SimilaritySearch.index! — Function

index!(index::SearchGraph; parallel_block=1, parallel_minimum_first_block=parallel_block, callbacks=SearchGraphCallbacks())

Indexes the already initialized database (e.g., given in the constructor method). It can be made in parallel or sequentially. The arguments are the same than append! function but using the internal index.db as input.

SimilaritySearch.rebuild — Function

rebuild(g::SearchGraph; neighborhood=Neighborhood(), callbacks=SearchGraphCallbacks(), minbatch=0, pools=getpools(index))

Rebuilds the SearchGraph index but seeing the whole dataset for the incremental construction, i.e., it can connect the i-th vertex to its knn in the 1..n possible vertices instead of its knn among 1..(i-1) as in the original algorithm.

Arguments

g: The search index to be rebuild.

neighborhood: The neighborhood strategy to follow in the rebuild, it can differ from the original one.

callbacks: The set of callbacks

minbatch: controls how the multithreading is made, see getminbatch

pools: The set of caches for the indexes

Distance functions

The distance functions are defined to work under the evaluate(::metric, u, v) function (borrowed from Distances.jl package).

Cosine and angle distances for vectors

SimilaritySearch.CosineDistance — Type

CosineDistance()

The cosine is defined as:

\[\cos(u, v) = \frac{\sum_i u_i v_i}{\sqrt{\sum_i u_i^2} \sqrt{\sum_i v_i^2}}\]

The cosine distance is defined as $1 - \cos(u,v)$

SimilaritySearch.NormalizedCosineDistance — Type

NormalizedCosineDistance()

Similar to CosineDistance but suppose that input vectors are already normalized, and therefore, reduced to simply one minus the dot product.

\[1 - \sum_i {u_i v_i}\]

SimilaritySearch.AngleDistance — Type

AngleDistance()

The angle distance is defined as:

\[∠(u, v)= \arccos(\cos(u, v))\]

SimilaritySearch.NormalizedAngleDistance — Type

NormalizedAngleDistance()

Similar to AngleDistance but suppose that input vectors are already normalized

\[\arccos \sum_i {u_i v_i}\]

Set distance functions

Set bbject are represented as ordered arrays

SimilaritySearch.JaccardDistance — Type

JaccardDistance()

The Jaccard distance is defined as

\[J(u, v) = \frac{|u \cap v|}{|u \cup v|}\]

SimilaritySearch.DiceDistance — Type

DiceDistance()

The Dice distance is defined as

\[D(u, v) = \frac{2 |u \cap v|}{|u| + |v|}\]

SimilaritySearch.IntersectionDissimilarity — Type

IntersectionDissimilarity()

The intersection dissimilarity uses the size of the intersection as a mesuare of similarity as follows:

\[I(u, v) = 1 - \frac{|u \cap v|}{\max \{|u|, |v|\}}\]

SimilaritySearch.CosineDistanceSet — Type

CosineDistanceSet()

The cosine distance for very sparse binary vectors represented as sorted lists of positive integers where ones occur.

String alignment distances

The following uses strings/arrays as input, i.e., objects follow the array interface. A broader set of distances for strings can be found in the StringDistances.jl package.

SimilaritySearch.CommonPrefixDissimilarity — Type

CommonPrefixDissimilarity()

Uses the common prefix as a measure of dissimilarity between two strings

SimilaritySearch.GenericLevenshteinDistance — Type

GenericLevenshteinDistance(;icost, dcost, rcost)

The levenshtein distance measures the minimum number of edit operations to convert one string into another. The costs insertion icost, deletion cost dcost, and replace cost rcost. Not thread safe, use a copy of for each thread.

SimilaritySearch.StringHammingDistance — Type

StringHammingDistance()

The hamming distance counts the differences between two equally sized strings

SimilaritySearch.LevenshteinDistance — Function

LevenshteinDistance()

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.LcsDistance — Function

LcsDistance()

Instantiates a GenericLevenshteinDistance object to perform LCS distance

Distances for Cloud of points

SimilaritySearch.HausdorffDistance — Type

HausdorffDistance(dist::SemiMetric)

Hausdorff distance is defined as the maximum of the minimum between two clouds of points.

\[Hausdorff(U, V) = \max{\max_{u \in U} nndist(u, V), \max{v \in V} nndist(v, U) }\]

where $nndist(u, V)$ computes the distance of $u$ to its nearest neighbor in $V$ using the dist metric.

SimilaritySearch.MinHausdorffDistance — Type

MinHausdorffDistance(dist::SemiMetric)

Similar to HausdorffDistance but using minimum instead of maximum.

Public API

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors.

Distances.evaluate — Method

evaluate(::BinaryHammingDistance, a, b)
evaluate(::BinaryHammingDistance, a::AbstractVector, b::AbstractVector) where {T<:Unsigned}

Computes the binary hamming distance for bit types and arrays of bit types

Distances.evaluate — Method

evaluate(::CommonPrefixDissimilarity, a, b)

Computes a dissimilarity based on the common prefix between two strings

Distances.evaluate — Method

evaluate(::CosineDistance, a, b)

Computes the cosine distance between two vectors. Please use AngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::CosineDistanceSet, a, b)

Computes the cosine distance where a and b are sorted lists of integers (emulating binary sparse vectores)

Distances.evaluate — Method

evaluate(::DiceDistance, a, b)

Computes the Dice's distance of a and b both sets specified as sorted vectors.

Distances.evaluate — Method

evaluate(::GenericLevenshteinDistance, a, b)

Computes the edit distance between two strings, this is a low level function

Distances.evaluate — Method

evaluate(m::HausdorffDistance, u, v)

Computes the Hausdorff distance between two cloud of points.

u and v are iterables where each object can be measured with the internal distance dist

Distances.evaluate — Method

evaluate(::IntersectionDissimilarity, a, b)

Uses the intersection as a distance function (non-metric)

Distances.evaluate — Method

evaluate(::JaccardDistance, a, b)

Computes the Jaccard's distance of a and b both sets specified as sorted vectors.

Distances.evaluate — Method

evaluate(::L1Distance, a, b)

Computes the Manhattan's distance between a and b

Distances.evaluate — Method

evaluate(::L2Distance, a, b)

Computes the Euclidean's distance betweem a and b

Distances.evaluate — Method

evaluate(::LInftyDistance, a, b)

Computes the maximum distance or Chebyshev's distance

Distances.evaluate — Method

evaluate(lp::LpDistance, a, b)

Computes generic Minkowski's distance

Distances.evaluate — Method

evaluate(m::MinHausdorffDistance, u, v)

Computes a variant of the Hausdorff distance that uses the minimum instead of the maximum. u and v are iterables where each object can be measured with the internal distance dist

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors. It supposes that all vectors are normalized

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a, b)

Computes the cosine distance between two vectors, it expects normalized vectors. Please use NormalizedAngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Distances.evaluate — Method

evaluate(::SqL2Distance, a, b)

Computes the squared Euclidean's distance between a and b

Distances.evaluate — Method

 evaluate(::StringHammingDistance, a, b)

Computes the hamming distance between two sequences of the same length

SimilaritySearch.LcsDistance — Method

LcsDistance()

Instantiates a GenericLevenshteinDistance object to perform LCS distance

SimilaritySearch.LevenshteinDistance — Method

LevenshteinDistance()

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

SimilaritySearch.allknn — Method

allknn(g, knns, dists; parallel_block=512, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Arguments:

g: the index

knns: an uninitialized integer matrix of (k, n) size for storing the k nearest neighbors of the n elements

dists: an uninitialized floating point matrix of (k, n) size for storing the k nearest distances of the n elements

minbatch: controls how multithreading is used for evaluating configurations, see getminbatch

pools: A pools object, dependent of g

Results:

knns and dists are returned. Note that the index can retrieve less than k objects, and these are represented as zeros at the end of each column (can happen)

SimilaritySearch.allknn — Method

allknn(g::AbstractSearchContext, k::Integer; minbatch=0, pools=getpools(g)) -> knns, dists

Computes all the k nearest neighbors (all vs all) using the index g. It removes self references.

Parameters:

g: the index

k: the number of neighbors to retrieve

minbatch: controls how multithreading is used for evaluating configurations, see getminbatch

pools: A pools object, dependent of g

Returns:

knns a (k, n) matrix of identifiers; the i-th column corresponds to the i-th object in the dataset. Zeros can happen to the end of each column meaning that the retrieval was less than the desired k

dists a (k, n) matrix of distances; the i-th column corresponds to the i-th object in the dataset. Zero values in knns should be ignored in dists

SimilaritySearch.closestpair — Method

closestpair(idx::AbstractSearchContext; parallel=false, pools=getpools(idx))

Finds the closest pair among all elements in idx. If the index idx is approximate then pair of points could be also an approximation.

Arguments:

idx: the search structure that indexes the set of points

Keyword Arguments:

parallel: If true then the algorithm uses all available threads to compute the closest pair

pools: The pools needed for the index. Only used for special cases, default values should work in most cases. See getpools for more information.

SimilaritySearch.find_neighborhood — Method

find_neighborhood(index::SearchGraph{T}, item, neighborhood, pools; hints=index.hints)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function). res is always reused since reduce creates a new KnnResult from it (a copy if reduce in its simpler terms)

Arguments

index: The search index.

item: The item to be inserted.

neighborhood: A Neighborhood object that describes how to compute item's neighborhood.

pools: Cache pools to be used

hints: Search hints

SimilaritySearch.getknnresult — Function

getknnresult(k::Integer, pools=nothing) -> KnnResult

Generic function to obtain a shared result set for the same thread and avoid memory allocations. This function should be specialized for indexes and pools that use shared results or threads in some special way.

SimilaritySearch.index! — Method

index!(index::SearchGraph; parallel_block=1, parallel_minimum_first_block=parallel_block, callbacks=SearchGraphCallbacks())

Indexes the already initialized database (e.g., given in the constructor method). It can be made in parallel or sequentially. The arguments are the same than append! function but using the internal index.db as input.

SimilaritySearch.macrorecall — Function

macrorecall(goldI::AbstractMatrix, resI::AbstractMatrix, k=size(goldI, 1))::Float64

Computes the macro recall score using goldI as gold standard and resI as predictions; it expects that matrices of integers (identifiers). If k is given, then the results are cut to first k.

SimilaritySearch.macrorecall — Method

macrorecall(goldlist::AbstractVector, reslist::AbstractVector)::Float64

Computes the macro recall score using sets of results (KnnResult objects or vectors of itegers).

SimilaritySearch.maxlength — Method

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

SimilaritySearch.neardup — Method

neardup(idx::AbstractSearchContext, X::AbstractVector, ϵ)

Find nearest duplicates in database X using the empty index idx. The algorithm iteratively try to index elements in X, and items being near than ϵ to some element in idx will be ignored.

The function returns a named tuple (idx, map, nn, dist) where:

idx: it is the index of the non duplicated elements

map: a mapping from 1-|idx| to its positions in X

nn: an array where each element in $x \in X$ points to its covering element (previously indexed element u such that $d(u, x_i) \leq ϵ$)

dist: an array of distance values to each covering element (correspond to each element in nn)

Arguments

idx: An empty index (i.e., a SearchGraph)

X: The input dataset

ϵ: Real value to cut

Notes

The index idx must support incremental construction with push!

You can access the set of elements being 'ϵ'-non duplicates (the $ϵ-net$) using idx.db or where nn[i] == i

SimilaritySearch.optimize! — Function

optimize!(
    index::AbstractSearchContext,
    kind::ErrorFunction=ParetoRecall(),
    space::AbstractSolutionSpace=optimization_space(index);
    queries=nothing,
    ksearch=10,
    numqueries=64,
    initialpopulation=8,
    minbatch=0,
    verbose=false,
    params=SearchParams(; maxpopulation=8, bsize=4, mutbsize=8, crossbsize=2, tol=-1.0, maxiters=8, verbose)
)

Tries to configure the index to achieve the specified performance (kind). The optimization procedure is an stochastic search over the configuration space yielded by kind and queries.

Arguments

index: the index to be optimized

kind: The kind of optimization to apply, it can be ParetoRecall(), ParetoRadius() or MinRecall(r) where r is the expected recall (0-1, 1 being the best quality but at cost of the search time)

Keyword arguments

queries: the set of queries to be used to measure performances, a validation set. It can be an AbstractDatabase or nothing.

queries_ksearch: the number of neighbors to retrieve for queries

queries_size: if queries===nothing then a sample of the already indexed database is used, queries_size is the size of the sample.

initialpopulation: the initial sample for the optimization procedure

minbatch: controls how multithreading is used for evaluating configurations, see getminbatch

space: defines the search space

params: the parameters of the solver, see search_models function from SearchModels.jl package for more information.

verbose: controls if the procedure is verbose or not

SimilaritySearch.prune! — Method

prune!(r::KeepNearestPruning, index::SearchGraph; pools=getpools(index))

Selects k nearest neighbors among the the available neighbors

SimilaritySearch.prune! — Method

prune!(r::RandomPruning, index::SearchGraph;  minbatch=0, pools=getpools(index))

Randomly prunes each neighborhood

Arguments

SimilaritySearch.prune! — Method

prune!(r::SatPruning, index::SearchGraph; pools=getpools(index))

Select the SatNeighborhood or DistalSatNeighborhood from available neihghbors

SimilaritySearch.push_item! — Method

push_item!(
    index::SearchGraph,
    item,
    neighborhood,
    push_item,
    callbacks,
    pools
)

Appends an object into the index

Arguments:

index: The search graph index where the insertion is going to happen.

item: The object to be inserted, it should be in the same space than other objects in the index and understood by the distance metric.

neighborhood: A Neighborhood object that specifies the kind of neighborhood that will be computed.

push_item: if false is an internal option, used by append! and index! (it avoids to insert item into the database since it is already inserted but not indexed).

callbacks: The set of callbacks that are called whenever the index grows enough. Keeps hyperparameters and structure in shape.

pools: The set of caches used for searching.

Note: setting callbacks as nothing ignores the execution of any callback

SimilaritySearch.push_neighborhood! — Method

push_neighborhood!(index::SearchGraph, item, neighbors, callbacks; push_item=true)

Inserts the object item into the index, i.e., creates an edge for each item in neighbors (internal function)

Arguments

index: The search index to be modified.

item: The item that will be inserted.

neighbors: An array of indices that will be connected to the new vertex.

callbacks: A [SearchGraphCallbacks] object (callback list) that will be called after some insertions

push_item: Specifies if the item must be inserted into the internal db (sometimes is already there like in index!)

SimilaritySearch.rebuild — Method

rebuild(g::SearchGraph; neighborhood=Neighborhood(), callbacks=SearchGraphCallbacks(), minbatch=0, pools=getpools(index))

Rebuilds the SearchGraph index but seeing the whole dataset for the incremental construction, i.e., it can connect the i-th vertex to its knn in the 1..n possible vertices instead of its knn among 1..(i-1) as in the original algorithm.

Arguments

g: The search index to be rebuild.

neighborhood: The neighborhood strategy to follow in the rebuild, it can differ from the original one.

callbacks: The set of callbacks

minbatch: controls how the multithreading is made, see getminbatch

pools: The set of caches for the indexes

SimilaritySearch.recallscore — Method

recallscore(gold::Set, res)

Compute recall and precision scores from the result sets.

SimilaritySearch.reuse! — Function

reuse!(res::KnnResult)
reuse!(res::KnnResult, k::Integer)

Returns a result set and a new initial state; reuse the memory buffers

SimilaritySearch.search — Method

search(bs::BeamSearch, index::SearchGraph, q, res, hints, pools; bsize=bs.bsize, Δ=bs.Δ, maxvisits=bs.maxvisits)

Tries to reach the set of nearest neighbors specified in res for q.

bs: the parameters of BeamSearch

index: the local search index

q: the query

res: The result object, it stores the results and also specifies the kind of query

hints: Starting points for searching, randomly selected when it is an empty collection

pools: A SearchGraphPools object with preallocated pools

Optional arguments (defaults to values in bs)

bsize: Beam size

Δ: exploration expansion factor

maxvisits: Maximum number of nodes to visit (distance evaluations)

SimilaritySearch.search — Method

search(seq::ExhaustiveSearch, q, res::KnnResult)

Solves the query evaluating all items in the given query.

SimilaritySearch.search — Method

search(ex::ParallelExhaustiveSearch, q, res::KnnResult; minbatch=0, pools=nothing)

Solves the query evaluating all items in the given query.

Arguments

ex: the search structure

q: the query to solve

res: the result set

Keyword arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: The set of caches (nothing for this index)

SimilaritySearch.search — Method

search(index::SearchGraph, q, res; hints=index.hints, pools=getpools(index))

Solves the specified query res for the query object q.

SimilaritySearch.searchbatch — Method

searchbatch(index, Q, I::AbstractMatrix{Int32}, D::AbstractMatrix{Float32}; minbatch=0, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index and I and D as output (searches for k=size(I, 1))

Arguments

index: The search structure

Q: The set of queries

k: The number of neighbors to retrieve

Keyword arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

SimilaritySearch.searchbatch — Method

searchbatch(index, Q, KNN::AbstractVector{KnnResult}; minbatch=0, pools=getpools(index)) -> indices, distances

Searches a batch of queries in the given index using an array of KnnResult's; each KnnResult object can specify different k values.

Arguments

minbatch: Minimum number of queries solved per each thread, see getminibatch

pools: relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

SimilaritySearch.searchbatch — Method

searchbatch(index, Q, k::Integer; minbatch=0, pools=GlobalKnnResult) -> indices, distances

Searches a batch of queries in the given index (searches for k neighbors).

Arguments

index: The search structure

Q: The set of queries

k: The number of neighbors to retrieve

Keyword arguments

minbatch specifies how many queries are solved per thread.

Integers $1 ≤ minbatch ≤ |Q|$ are valid values

Set minbatch=0 to compute a default number based on the number of available cores.

Set minbatch=-1 to avoid parallelism.

pools relevant for special datasets or distance functions. In most case uses the default is enought, but different pools should be used when indexes use other indexes internally to solve queries. It should be an array of Threads.nthreads() preallocated KnnResult objects used to reduce memory allocations.

Note: The i-th column in indices and distances correspond to the i-th query in Q Note: The final indices at each column can be 0 if the search process was unable to retrieve k neighbors.

SimilaritySearch.timedsearchbatch — Method

timedsearchbatch(index, Q, ksearch::Integer; parallel=false)

Computes the K nearest neigbors of each object in Q and returns two matrices, and the average search time in seconds.

SimilaritySearch.AbstractDatabase — Type

abstract type AbstractDatabase end

Base type to represent databases. A database is a collection of objects that can be accessed like a similar interface to AbstractVector. It is separated to allow SimilaritySearch methods to know what is a database and what is an object (since most object representations will look as vectors and matrices).

The basic implementations are:

MatrixDatabase: A wrapper for object-vectors stored in a Matrix, columns are the objects. It is static.

DynamicMatrixDatabase: A dynamic representation for vectors that allows adding new vectors.

VectorDatabase: A wrapper for vector-like structures. It can contain any kind of objects.

SubDatabase: A sample of a given database

In particular, the storage details are not used by VectorDatabase and MatrixDatabase. For instance, it is possible to use matrices like Matrix, SMatrix or StrideArrays; or even use generated objects with VectorDatabase (supporting a vector-like interface).

If the storage backend support it, it is possible to use vector operations, for example:

get the i-th element obj = db[i], elements in the database are identified by position

get the elements list in a list of indices lst as db[lst] (also using view)

set a value at the i-th element db[i] = obj

random sampling rand(db), rand(db, 3)

iterate and collect objects in the database

get the number of elements in the database length(db)

add new objects to the end of the database (not all internal containers will support it)

push!(db, u) adds a single element u

append!(db, lst) adds a list of objects to the end of the database

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, Δ::Float32)

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time.

bsize: The size of the beam.

Δ: Soft margin for accepting elements into the beam

maxvisits: MAximum visits while searching, useful for early stopping without convergence

SimilaritySearch.BinaryHammingDistance — Type

BinaryHammingDistance()

Binary hamming uses bit wise operations to count the differences between bit strings

SimilaritySearch.Callback — Type

abstract type Callback end

Abstract type to trigger callbacks after some number of insertions. SearchGraph stores the callbacks in callbacks (a dictionary that associates symbols and callback objects); A SearchGraph object controls when callbacks are fired using callback_logbase and callback_starting

SimilaritySearch.DisjointHints — Type

mutable struct DisjointHints

Indicates that hints are a small disjoint (untouched neighbors) subsample

SimilaritySearch.DistalSatNeighborhood — Type

DistalSatNeighborhood()

New items are connected with a small set of items computed with a Distal SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage but in reverse order of distance.

SimilaritySearch.DynamicMatrixDatabase — Type

struct DynamicMatrixDatabase{DType,Dim} <: AbstractDatabase

A dynamic matrix with elements of type DType and dimension Dim

SimilaritySearch.DynamicMatrixDatabase — Method

DynamicMatrixDatabase(M::MatrixDatabase)

Creates a DynamicMatrixDatabase from a MatrixDatabase, copies internal data. Please see AbstractDatabase for general usage.

SimilaritySearch.DynamicMatrixDatabase — Method

DynamicMatrixDatabase(matrix::AbstractMatrix{DType})

Creates a DynamicMatrixDatabase from a matrix-like object. Please see AbstractDatabase for general usage.

SimilaritySearch.DynamicMatrixDatabase — Method

DynamicMatrixDatabase(::Type{DType}, Dim::Integer)

Creates an empty DynamicMatrixDatabase such that length(db[i]) == Dim and eltype(db[i]) == DType

SimilaritySearch.DynamicMatrixDatabase — Method

DynamicMatrixDatabase(V::DynamicMatrixDatabase{DType,Dim})

Creates a DynamicMatrixDatabase from another DynamicMatrixDatabase, they will share their internal data. Please see AbstractDatabase for general usage.

SimilaritySearch.IdentityNeighborhood — Type

struct IdentityNeighborhood

It does not modifies the given neighborhood

SimilaritySearch.KDisjointHints — Type

struct KDisjointHints

Indicates that hints are selected to have a disjoint neighborhood

SimilaritySearch.KeepNearestPruning — Type

KeepNearestPruning(k)

Kept k nearest neighbor edges for each vertex

SimilaritySearch.KnnResult — Type

KnnResult(ksearch::Integer)

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

SimilaritySearch.L1Distance — Type

L1Distance()

The manhattan distance or $L_1$ is defined as

\[L_1(u, v) = \sum_i{|u_i - v_i|}\]

SimilaritySearch.L2Distance — Type

L2Distance()

The euclidean distance or $L_2$ is defined as

\[L_2(u, v) = \sqrt{\sum_i{(u_i - v_i)^2}}\]

SimilaritySearch.LInftyDistance — Type

LInftyDistance()

The Chebyshev or $L_{\infty}$ distance is defined as

\[L_{\infty}(u, v) = \max_i{\left| u_i - v_i \right|}\]

SimilaritySearch.LpDistance — Type

LpDistance(p)
LpDistance(p, pinv)

The general Minkowski distance $L_p$ distance is defined as

\[L_p(u, v) = \left|\sum_i{(u_i - v_i)^p}\right|^{1/p}\]

Where $p_{inv} = 1/p$. Note that you can specify unrelated p and pinv if you need an specific behaviour.

SimilaritySearch.MatrixDatabase — Type

struct MatrixDatabase{M<:AbstractDatabase} <: AbstractDatabase

MatrixDatabase(matrix::AbstractMatrix)

Wraps a matrix-like object matrix into a MatrixDatabase. Please see AbstractDatabase for general usage.

SimilaritySearch.MatrixDatabase — Method

MatrixDatabase(V::MatrixDatabase)

Creates another MatrixDatabase from another MatrixDatabase. They will share their internal data. Please see AbstractDatabase for general usage.

SimilaritySearch.NegativeDistanceHack — Type

NegativeDistanceHack(dist)

Evaluates as the negative of the distance function being wrapped. This is not a real distance function but a simple hack to get a similarity and use it for searching for farthest elements (farthest points / farthest pairs) on indexes that can handle this hack (e.g., ExhaustiveSearch, ParallelExhaustiveSearch, SearchGraph).

SimilaritySearch.Neighborhood — Type

@with_kw mutable struct Neighborhood

Determines the size of the neighborhood, $k$ is adjusted as a callback, and it is intended to affect previously inserted vertices. The neighborhood is designed to consider two components $k=in+out$, i.e. incoming and outgoing edges for each vertex.

The $out$ size is computed as $minsize + \log(logbase, n)$ where $n$ is the current number of indexed elements; this is computed searching

for $out$ elements in the current index.

The $in$ size is unbounded.

reduce is intended to postprocess neighbors (after search process, i.e., once out edges are computed); do not change $k$ but always must return a copy of the reduced result set.

Note: Set $logbase=Inf$ to obtain a fixed number of $in$ nodes; and set $minsize=0$ to obtain a pure logarithmic growing neighborhood.

SimilaritySearch.Neighborhood — Method

Neighborhood(reduce::NeighborhoodReduction; logbase=2, minsize=2)

Convenience constructor, see Neighborhood struct.

SimilaritySearch.ParallelExhaustiveSearch — Method

ParallelExhaustiveSearch(; dist=SqL2Distance(), db=VectorDatabase{Float32}())

Solves queries evaluating dist in parallel for the query and all elements in the dataset. Note that this should not be used in conjunction with searchbatch(...; parallel=true) since they will compete for resources.

SimilaritySearch.RandomHints — Type

mutable struct RandomHints

Indicates that hints are a random sample of the dataset

SimilaritySearch.RandomPruning — Type

RandomPruning(k)

Selects k random edges for each vertex

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood()

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

SimilaritySearch.SatPruning — Type

SatPruning(k, satkind)
SatPruning(k)

Selects SatNeighborhood or DistalSatNeighborhood for each vertex. Defaults to DistalSatNeighborhood.

k: the threshold size to apply the Sat reduction, i.e., neighbors larger than k will be pruned.

SimilaritySearch.SearchGraphCallbacks — Type

struct SearchGraphCallbacks hints::Union{Nothing,Callback} = DisjointHints() hyperparameters::Union{Nothing,Callback} = OptimizeParameters(kind=ParetoRecall()) logbase::Float32 = 1.5 starting::Int32 = 8 end

Call insertions and indexing methods with SearchGraphCallbacks objects to control how the index structure is adjusted (callbacks are called when $n > starting$ and $\lceil(\log(logbase, n)\rceil != \lceil\log(logbase, n+1)\rceil$)

SimilaritySearch.SearchGraphCallbacks — Method

SearchGraphCallbacks(kind::ErrorFunction;
    hints=DisjointHints(),
    logbase=1.5,
    starting=8,
    initialpopulation=16,
    maxpopulation=initialpopulation,
    maxiters=12,
    bsize=4,
    mutbsize=4bsize,
    crossbsize=2bsize,
    tol=-1.0,
    params=SearchParams(; maxpopulation, bsize, mutbsize, crossbsize, tol, maxiters),
    ksearch=10,
    numqueries=32,
    space::BeamSearchSpace=BeamSearchSpace()
)

Convenient constructor function to create SearchGraphCallbacks structs. See SearchGraphCallbacks, SearchParams, and BeamSearchSpace for more details.

Arguments

kind: The kind of error function, e.g. MinRecall(0.9).

hints: How search hints should be computed.

logbase: Controls the periodicity of callback executions.

starting: Controls the minimum size of the index before execute callbacks.

initialpopulation: Optimization argument that determines the initial number of configurations.

maxiters: Optimization argument that determines the number of iterations.

bsize: Optimization argument that determines how many top configurations are allowed to mutate and cross.

tol: Optimization argument that determines the minimal tolerance improvement to stop the optimization.

params: The SearchParams arguments (if separated optimization arguments are not enough)

ksearch: The number of neighbors to be retrived by the optimization process.

numqueries: The number of queries to be performed during the optimization process.

space: The cofiguration search space

SimilaritySearch.SearchGraphPools — Type

SearchGraphPools(results=GlobalKnnResult, vstates=GlobalVisitedVertices, beams=GlobalBeamKnnResult)

A set of pools to alleviate memory allocations in SearchGraph construction and searching. Relevant on multithreading scenarious where distance functions, evaluate can call other metric indexes that can use these shared resources (globally defined).

Each pool is a vector of Threads.nthreads() preallocated objects of the required type.

SimilaritySearch.SimilarityFromDistance — Type

SimilarityFromDistance(dist)

Evaluates as $1/(1 + d)$ for a distance evaluation $d$ of dist. This is not a distance function and is part of the hacks to get a similarity for searching farthest elements on indexes that can handle this hack (e.g., ExhaustiveSearch, ParallelExhaustiveSearch, SearchGraph).

SimilaritySearch.SqL2Distance — Type

SqL2Distance()

The squared euclidean distance is defined as

\[L_2(u, v) = \sum_i{(u_i - v_i)^2}\]

It avoids the computation of the square root and should be used whenever you are able do it.

SimilaritySearch.VectorDatabase — Type

struct VectorDatabase{V} <: AbstractDatabase

A wrapper for vector-like databases

SimilaritySearch.VectorDatabase — Method

VectorDatabase(D::AbstractDatabase)

Creates a VectorDatabase from an AbstractDatabase. It copies internal data. Please see AbstractDatabase for general usage.

SimilaritySearch.VectorDatabase — Method

VectorDatabase(M::AbstractMatrix)

Creates a VectorDatabase from a matrix-like object. It will copy columns as objects of the new VectorDatabase. Please see AbstractDatabase for general usage.

SimilaritySearch.VectorDatabase — Method

VectorDatabase(vecs::T)

Creates a VectorDatabase from vecs. Please see AbstractDatabase for general usage.

SimilaritySearch.VectorDatabase — Method

VectorDatabase(V::VectorDatabase)

Creates a VectorDatabase from another an AbstractDatabase. They will share their internal data. Please see AbstractDatabase for general usage.

SimilaritySearch.VectorDatabase — Method

VectorDatabase(; type=Vector{Float32})

Creates an empty VectorDatabase where each object is of type type. Please see AbstractDatabase for general usage.

Private API

Base.append! — Method

append!(
    index::SearchGraph,
    db;
    neighborhood=Neighborhood(),
    parallel_block=1,
    parallel_minimum_first_block=parallel_block,
    callbacks=SearchGraphCallbacks(),
    pools=getpools(index)
)

Appends all items in db to the index. It can be made in parallel or sequentially.

Arguments:

index: the search graph index

neighborhood: A Neighborhood object that specifies the kind of neighborhood that will be computed.

db: the collection of objects to insert, an AbstractDatabase is the canonical input, but supports any iterable objects

parallel_block: The number of elements that the multithreading algorithm process at once, it is important to be larger that the number of available threads but not so large since the quality of the search graph could degrade (a few times the number of threads is enough). If parallel_block=1 the algorithm becomes sequential.

parallel_minimum_first_block: The number of sequential appends before running parallel.

Note: Parallel doesn't trigger callbacks inside blocks.

callbacks: A SearchGraphCallbacks object to be called after some insertions (specified by the callbacks object). These callbacks are used to maintain the algorithm in good shape after many insertions (adjust hyperparameters and the structure).

pools: The set of caches used for searching.

Note 1: Callbacks are not executed inside parallel blocks Note 2: Callbacks will be ignored if callbacks=nothing

Base.convert — Method

convert(::Type{AbstractDatabase}, M::AbstractDatabase)
convert(::Type{AbstractDatabase}, M::AbstractMatrix)
convert(::Type{AbstractDatabase}, M::Vector)
convert(::Type{AbstractDatabase}, M::Vector{Any})
convert(::Type{AbstractDatabase}, M::AbstractVector)
convert(::Type{<:AbstractVector}, M::VectorDatabase)
convert(::Type{<:AbstractVector}, M::AbstractDatabase)

Convenience functions to convert different kinds of data into a some kind of database.

Base.eachindex — Method

eachindex(db::AbstractDatabase)

An index iterator of db

Base.eltype — Method

eltype(db::DynamicMatrixDatabase{DType,Dim})

The type of stored elements AbstractVector{DType}

Base.firstindex — Method

firstindex(db::AbstractDatabase)

First index in db (1)

Base.getindex — Method

getindex(db::AbstractDatabase, lst)

A subset of db (using a collection of indexes lst)

Base.getindex — Method

getindex(res::KnnResult, i)

Access the i-th item in res

Base.iterate — Function

Base.iterate(res::KnnResult, state::Int=1)

Support for iteration

Base.lastindex — Method

lastindex(db::AbstractDatabase)

Last index in db (1)

Base.pop! — Method

pop!(res::KnnResult)

Removes and returns the last item in the pool, it is an O(1) operation

Base.popfirst! — Method

popfirst!(p::KnnResult)

Removes and returns the nearest neeighboor pair from the pool, an O(length(p.pool)) operation

Base.push! — Method

push!(res::KnnResult, item::Pair)
push!(res::KnnResult, id::Integer, dist::Real)

Appends an item into the result set

Base.rand — Method

rand(db::AbstractDatabase, n)

Retrieves n random elements from db, returnes a SubDatabase object

Base.rand — Method

rand(db::AbstractDatabase)

Retrieves a random element from db

Base.size — Method

size(db::AbstractDatabase)

The size of the database (length(db),). There is no concept of explicit dimension for AbstractDatabase implementations.

Base.view — Method

view(db::AbstractDatabase, map)

Constructs a SubDatabase from db using the specified indexes in map (e.g., an array or an slice of indices)

SimilaritySearch._shifted_fixorder! — Method

_shifted_fixorder!(res::KnnResult, sp, ep)

Sorts the result in place; the possible element out of order is on the last entry always. It implements a kind of insertion sort that it is efficient due to the expected distribution of the items being inserted (it is expected just a few elements smaller than the current ones)

SimilaritySearch.common_prefix — Method

common_prefix(a, b)

Computes the length of the common prefix among two strings represented as arrays

SimilaritySearch.execute_callback — Method

execute_callback(opt::KDisjointHints, index)

SearchGraph's callback for selecting hints at random

SimilaritySearch.execute_callback — Method

execute_callback(opt::OptimizeParameters, index::SearchGraph)

SearchGraph's callback for adjunting search parameters

SimilaritySearch.execute_callback — Method

executed_callback(opt::RandomHints, index)

SearchGraph's callback for selecting hints at random

SimilaritySearch.execute_callbacks — Function

execute_callbacks(index::SearchGraph, n=length(index), m=n+1)

Process all registered callbacks

SimilaritySearch.getminibatch — Method

getminibatch(minbatch, n)

Used by functions that use parallelism based on Polyester.jl minibatches specify how many queries (or something else) are solved per thread whenever the thread is used (in minibatches).

Arguments

minbatch

Integers $1 ≤ minbatch ≤ n$ are valid values (where n is the number of objects to process, i.e., queries)

Defaults to 0 which computes a default number based on the number of available cores.

Set minbatch=-1 to avoid parallelism.

SimilaritySearch.getpools — Method

getpools(index::SearchGraph)

Creates or retrieve caches for the search graph.

SimilaritySearch.intersectionsize — Function

intersectionsize(a, b, o=Forward)

Computes the size the intersections of a and b, specified as ordered sequences.

SimilaritySearch.neighborhoodreduce — Function

reduce(sat::DistalSatNeighborhood, index::SearchGraph, item, res, pools)

Reduces res using the DistSAT strategy.

SimilaritySearch.unionsize — Method

unionsize(a, b, isize)

Computes the size of the union of a and b that have an intersection size isize

SimilaritySearch.NeighborhoodPruning — Type

NeighborhoodPruning

Abstract data type for neighborhood pruning strategies

SimilaritySearch.NeighborhoodReduction — Type

abstract type NeighborhoodReduction end

Overrides Base.reduce(::NeighborhoodReduction, res::KnnResult, index::SearchGraph) to postprocess res using some criteria. Called from find_neighborhood, and returns a new KnnResult struct (perhaps a copy of res) since push_neighborhood captures the reference of its output.