Cosine distance functions

SimilaritySearch implements some cosine/angle distance functions. Please recall that we use the evaluate function definition from Distances.jl.

Cosine distance function

SimilaritySearch.CosineDistance — Type

CosineDistance()

The cosine is defined as:

\[\cos(u, v) = \frac{\sum_i u_i v_i}{\sqrt{\sum_i u_i^2} \sqrt{\sum_i v_i^2}}\]

The cosine distance is defined as $1 - \cos(u,v)$

Distances.evaluate — Method

evaluate(::CosineDistance, a, b)

Computes the cosine distance between two vectors. Please use AngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Angle distance function

SimilaritySearch.AngleDistance — Type

AngleDistance()

The angle distance is defined as:

\[∠(u, v)= \arccos(\cos(u, v))\]

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors.

Normalized cosine distance function

SimilaritySearch.NormalizedCosineDistance — Type

NormalizedCosineDistance()

Similar to CosineDistance but suppose that input vectors are already normalized

\[1 - \sum_i {u_i v_i}\]

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a, b)

Computes the cosine distance between two vectors, it expects normalized vectors (see normalize! method). Please use NormalizedAngleDistance if you are expecting a metric function (cosine_distance is a faster alternative whenever the triangle inequality is not needed)

Normalized angle distance function

SimilaritySearch.NormalizedAngleDistance — Type

NormalizedAngleDistance()

Similar to AngleDistance but suppose that input vectors are already normalized

\[\arccos \sum_i {u_i v_i}\]

Distances.evaluate — Method

evaluate(::AngleDistance, a, b)

Computes the angle between twovectors. It supposes that all vectors are normalized (see normalize! function)