Hamming distance functions

The hamming distance for binary and string data is implemented. Please recall that we use the evaluate function definition from Distances.jl.

Binary hamming distance function

SimilaritySearch.BinaryHammingDistance — Type

BinaryHammingDistance()

Binary hamming uses bit wise operations to count the differences between bit strings

Distances.evaluate — Method

evaluate(::BinaryHammingDistance, a, b)
evaluate(::BinaryHammingDistance, a::AbstractVector, b::AbstractVector) where {T<:Unsigned}

Computes the binary hamming distance for bit types and arrays of bit types

Hamming distance function

SimilaritySearch.StringHammingDistance — Type

StringHammingDistance()

The hamming distance counts the differences between two equally sized strings

Distances.evaluate — Method

 evaluate(::StringHammingDistance, a, b)

Computes the hamming distance between two sequences of the same length