Minkowski family of distance functions

The following distinguished members of the Minkowski family of distance functions are provided; please recall that we use the evaluate function definition from Distances.jl.

SimilaritySearch.L1Distance — Type

L1Distance()

The manhattan distance or $L_1$ is defined as

\[L_1(u, v) = \sum_i{|u_i - v_i|}\]

Distances.evaluate — Method

evaluate(::L1Distance, a, b)

Computes the Manhattan's distance between a and b

Euclidean distance

SimilaritySearch.L2Distance — Type

L2Distance()

The euclidean distance or $L_2$ is defined as

\[L_2(u, v) = \sqrt{\sum_i{(u_i - v_i)^2}}\]

Distances.evaluate — Method

evaluate(::L2Distance, a, b)

Computes the Euclidean's distance betweem a and b

Squared euclidean distance

SimilaritySearch.SqL2Distance — Type

SqL2Distance()

The squared euclidean distance is defined as

\[L_2(u, v) = \sum_i{(u_i - v_i)^2}\]

It avoids the computation of the square root and should be used whenever you are able do it.

Distances.evaluate — Method

evaluate(::SqL2Distance, a, b)

Computes the squared Euclidean's distance between a and b

Chebyshev distance

SimilaritySearch.LInftyDistance — Type

LInftyDistance()

The Chebyshev or $L_{\infty}$ distance is defined as

\[L_{\infty}(u, v) = \max_i{\left| u_i - v_i \right|}\]

Distances.evaluate — Method

evaluate(::LInftyDistance, a, b)

Computes the maximum distance or Chebyshev's distance

Generic Minkowski distance functions

SimilaritySearch.LpDistance — Type

LpDistance(p)
LpDistance(p, pinv)

The general Minkowski distance $L_p$ distance is defined as

\[L_p(u, v) = \left|\sum_i{(u_i - v_i)^p}\right|^{1/p}\]

Where $p_{inv} = 1/p$. Note that you can specify unrelated p and pinv if you need an specific behaviour.

Distances.evaluate — Method

evaluate(lp::LpDistance, a, b)

Computes generic Minkowski's distance