Set distances

This functions use sets measure the distance between them; each set is represented as an array of sorted integers. Please recall that we use the evaluate function definition from Distances.jl. The following set distances are supported.

Jaccard distance function

SimilaritySearch.JaccardDistance — Type

JaccardDistance()

The Jaccard distance is defined as

\[J(u, v) = \frac{|u \cap v|}{|u \cup v|}\]

Distances.evaluate — Method

evaluate(::JaccardDistance, a, b)

Computes the Jaccard's distance of a and b both sets specified as sorted vectors.

Dice distance function

SimilaritySearch.DiceDistance — Type

DiceDistance()

The Dice distance is defined as

\[D(u, v) = \frac{2 |u \cap v|}{|u| + |v|}\]

Distances.evaluate — Method

evaluate(::DiceDistance, a, b)

Computes the Dice's distance of a and b both sets specified as sorted vectors.

Intersection dissimilarity function

SimilaritySearch.IntersectionDissimilarity — Type

IntersectionDissimilarity()

The intersection dissimilarity uses the size of the intersection as a mesuare of similarity as follows:

\[I(u, v) = 1 - \frac{|u \cap v|}{\max \{|u|, |v|\}}\]

Distances.evaluate — Method

evaluate(::IntersectionDissimilarity, a, b)

Uses the intersection as a distance function (non-metric)