String distances

The following string distances are supported. Please recall that we use the evaluate function definition from Distances.jl.

Levenshtein (edit) distance function

SimilaritySearch.LevenshteinDistance — Function

LevenshteinDistance()

Instantiates a GenericLevenshteinDistance object to perform traditional levenshtein distance

Missing docstring.

Missing docstring for evaluate(::LevenshteinDistance, a, b). Check Documenter's build log for details.

Longest common subsequence (LCS) distance function

SimilaritySearch.LcsDistance — Function

LcsDistance()

Instantiates a GenericLevenshteinDistance object to perform LCS distance

Missing docstring.

Missing docstring for evaluate(::LcsDistance, a, b). Check Documenter's build log for details.

Generic Levenshtein distance function

SimilaritySearch.GenericLevenshteinDistance — Type

GenericLevenshteinDistance(;icost, dcost, rcost)

The levenshtein distance measures the minimum number of edit operations to convert one string into another. The costs insertion icost, deletion cost dcost, and replace cost rcost. Not thread safe, use a copy of for each thread.

Distances.evaluate — Method

evaluate(::GenericLevenshteinDistance, a, b)

Computes the edit distance between two strings, this is a low level function

Common prefix dissimilarity function

SimilaritySearch.CommonPrefixDissimilarity — Type

CommonPrefixDissimilarity()

Uses the common prefix as a measure of dissimilarity between two strings

Distances.evaluate — Method

evaluate(::CommonPrefixDissimilarity, a, b)

Computes a dissimilarity based on the common prefix between two strings