KnnResult

SimilaritySearch's core is to solve knn searches; for this matter, it relies on a fixed size priority queue, the KnnResult struct and its related functions. The role of this struct is specifying and handling results.

Its usage without searches is pretty simple, as it is exemplified in the next code.

using SimilaritySearch

res = KnnResult(3)
for i in 1:3
    push!(res, i => rand())
end

println("first 3: ", collect(res))

for i in 4:10
    push!(res, i => rand())
end

println("final: ", collect(res))

Data structures

The KnnResult contains two aligned arrays 'id' and 'dist' and an indicator of its maximum capacity 'k'.

SimilaritySearch.KnnResult — Type

KnnResult(ksearch::Integer)

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

Length and capacity of the queue

using SimilaritySearch

res = KnnResult(3)
for i in 1:5
    push!(res, i => rand())
    println("current length: $(length(res)), capacity: $(maxlength(res)), first-dist: $(first(res).dist), last-dist: $(last(res).dist), covrad: $(covrad(res))")
end

The priority queue has a current length and a maximum capacity

Missing docstring.

Missing docstring for length(::KnnResult). Check Documenter's build log for details.

SimilaritySearch.maxlength — Function

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

Bounds of the queue

Missing docstring.

Missing docstring for covrad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for last(::KnnResult). Check Documenter's build log for details.

Missing docstring.

Missing docstring for first(::KnnResult). Check Documenter's build log for details.

Accessing and iterating

Missing docstring.

Missing docstring for getindex(::KnnResult, i). Check Documenter's build log for details.

Missing docstring.

Missing docstring for eachindex(::KnnResult). Check Documenter's build log for details.

Missing docstring.

Missing docstring for iterate(::KnnResult). Check Documenter's build log for details.

Adding and removing items

Missing docstring.

Missing docstring for empty!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(::KnnResult, id, dist). Check Documenter's build log for details.

Missing docstring.

Missing docstring for popfirst!(::KnnResult). Check Documenter's build log for details.

Missing docstring.

Missing docstring for pop!(::KnnResult). Check Documenter's build log for details.