KnnResult

SimilaritySearch's core is to solve knn searches; for this matter, it relies on a fixed size priority queue, the KnnResult struct and its related functions. The role of this struct is specifying and handling results.

Its usage without searches is pretty simple, as it is exemplified in the next code.

using SimilaritySearch

res = KnnResult(3)
for i in 1:3
    push!(res, i => rand())
end

println("first 3: ", collect(res))

for i in 4:10
    push!(res, i => rand())
end

println("final: ", collect(res))

first 3: Pair{Int32, Float32}[3 => 0.014443064, 2 => 0.062594354, 1 => 0.40106893]
final: Pair{Int32, Float32}[3 => 0.014443064, 2 => 0.062594354, 10 => 0.12642445]

Data structures

The KnnResult contains two aligned arrays 'id' and 'dist' and an indicator of its maximum capacity 'k'.

SimilaritySearch.KnnResult — Type

KnnResult(ksearch::Integer)

Creates a priority queue with fixed capacity (ksearch) representing a knn result set. It starts with zero items and grows with push!(res, id, dist) calls until ksearch size is reached. After this only the smallest items based on distance are preserved.

Length and capacity of the queue

using SimilaritySearch

res = KnnResult(3)
for i in 1:5
    push!(res, i => rand())
    println("current length: $(length(res)), capacity: $(maxlength(res)), first-dist: $(first(res).dist), last-dist: $(last(res).dist), covrad: $(covrad(res))")
end

The priority queue has a current length and a maximum capacity

Missing docstring.

Missing docstring for length(::KnnResult). Check Documenter's build log for details.

SimilaritySearch.maxlength — Function

maxlength(res::KnnResult)

The maximum allowed cardinality (the k of knn)

Bounds of the queue

SimilaritySearch.covrad — Function

covrad(p::KnnResult)

Returns the coverage radius of the result set; if length(p) < K then typemax(Float32) is returned

Missing docstring.

Missing docstring for last(::KnnResult). Check Documenter's build log for details.

Missing docstring.

Missing docstring for first(::KnnResult). Check Documenter's build log for details.

Accessing and iterating

Base.getindex — Method

getindex(res::KnnResult, i)

Access the i-th item in res

Base.eachindex — Method

eachindex(res::KnnResult)

Iterator of valid item indexes in res

Base.iterate — Method

Base.iterate(res::KnnResult, state::Int=1)

Support for iteration

Adding and removing items

Base.empty! — Function

empty!(res::KnnResult)
empty!(res::KnnResult, k::Integer)

Clears the content of the result pool. If k is given then the size of the pool is changed; the internal buffer is adjusted as needed (only grows).

Missing docstring.

Missing docstring for push!(::KnnResult, id, dist). Check Documenter's build log for details.

Base.popfirst! — Method

popfirst!(p::KnnResult)

Removes and returns the nearest neeighboor pair from the pool, an O(length(p.pool)) operation

Base.pop! — Method

pop!(p)

Removes and returns the last item in the pool, it is an O(1) operation