SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(L2Distance(), X, verbose=true)

for q in Q
    println(search(index, q, KnnResult(3)))
end

{SearchGraph: dist=L2Distance(), n=0, search_algo={BeamSearch: bsize=16, ssize=16, hints=0, neighborhood_algo=LogNeighborhood, knn=10}
added n=10000, neighborhood=14, {BeamSearch: bsize=16, ssize=16, hints=0, LogNeighborhood, 2021-02-23T18:28:43.787
KnnResult(3, 3, Item[Item(6672, 0.1089371f0), Item(6397, 0.17763191f0), Item(2383, 0.21175896f0)])
KnnResult(3, 3, Item[Item(9657, 0.05567341f0), Item(5159, 0.07233492f0), Item(3409, 0.08445235f0)])
KnnResult(3, 3, Item[Item(413, 0.09006916f0), Item(3207, 0.119104184f0), Item(7005, 0.1590054f0)])
KnnResult(3, 3, Item[Item(3560, 0.020573186f0), Item(9690, 0.0798248f0), Item(7405, 0.1653042f0)])
KnnResult(3, 3, Item[Item(405, 0.09813908f0), Item(8028, 0.10420397f0), Item(3040, 0.1259042f0)])

There are a several options for the construction of the index.

Related API

SimilaritySearch.SearchGraph — Type

SearchGraph(dist::PreMetric,
    db::AbstractVector;
    search_algo::LocalSearchAlgorithm=BeamSearch(),
    neighborhood_algo::NeighborhoodAlgorithm=LogNeighborhood(),
    automatic_optimization=false,
    recall=0.9,
    ksearch=10,
    tol=0.001,
    verbose=true)

Creates a SearchGraph object, i.e., an index to perform approximate search on db using the given search and neighbohood strategies. If automatic_optimization is true, then the structure tries to reach the given recall under the given ksearch.

Missing docstring.

Missing docstring for SearchGraphOptions. Check Documenter's build log for details.

The index can be created incrementally,

using SimilaritySearch

index = SearchGraph(L2Distance(), Vector{Float32}[])

for i in 1:1000
    push!(index, rand(Float32, 4))
end

search(index, rand(Float32, 4), 3)

KnnResult(3, 3, Item[Item(271, 0.116023935f0), Item(428, 0.1755375f0), Item(160, 0.18470258f0), Item(610, 0.16544613f0), Item(741, 0.17229928f0), Item(72, 0.17522754f0), Item(678, 0.18004507f0), Item(235, 0.19969332f0), Item(792, 0.20085838f0), Item(487, 0.2063139f0)])

The index construction can be customized with find_neighborhood and push_neighborhood

Missing docstring.

Missing docstring for push!(::SearchGraph, elem). Check Documenter's build log for details.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(algo::FixedNeighborhood, index::SearchGraph, item)

Finds a list of neighbors using the FixedNeighborhood criterion of item in the index

find_neighborhood(index::SearchGraph{T}, item)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, L::AbstractVector{Int32})

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

SimilaritySearch.optimize! — Function

function optimize!(search_algo::LocalSearchAlgorithm,
                   index::SearchGraph{T},
                   recall::Float64,
                   perf::Performance;
                   bsize::Int=4,
                   tol::Float64=0.01,
                   probes::Int=0) where T

Optimizes a local search index for an specific algorithm to get the desired performance. Note that optimizing for low-recall will yield to faster searches; the train queries are specified as part of the perf struct.

also see search and push!