SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(; dist=L2Distance(), verbose=true)
append!(index, X)

for q in Q
    println(search(index, q, KnnResult(3)))
end

calling callback optimize_hints; n=15
calling callback optimize_parameters; n=15
==== BEGIN Opt. {BeamSearch: bsize=8, hints=6, expected recall: 0.9, n: 15
  == Begin Opt. {BeamSearch: bsize=8, hints=6 iteration: 1, expected recall: 0.9, n: 15
    ** new best conf: (state={BeamSearch: bsize=9, hints=0, score=2.026315789473684, perf=StatsComparison(1.0, 1.0, 1.0, 2.1641071026141826e-6, 14.615384615384615, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), #beam: 2, n: 15
    ** new best conf: (state={BeamSearch: bsize=13, hints=0, score=2.089385474860335, perf=StatsComparison(0.944055944055944, 0.9496503496503497, 0.9467199467199467, 2.6226043701171875e-6, 13.76923076923077, StatsKnn(19.625336940471943, 1.0716500007189238, 2.839082360267639, 9.923076923076923), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), #beam: 3, n: 15
    ** new best conf: (state={BeamSearch: bsize=17, hints=0, score=2.0540540540540544, perf=StatsComparison(0.979020979020979, 0.979020979020979, 0.979020979020979, 2.6042644794170674e-6, 14.23076923076923, StatsKnn(19.600750336280235, 1.0716500007189238, 2.7964179607538076, 10.0), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), #beam: 4, n: 15
  == Iteration finished, current best: (state={BeamSearch: bsize=13, hints=0, score=2.089385474860335, perf=StatsComparison(0.944055944055944, 0.9496503496503497, 0.9467199467199467, 2.6226043701171875e-6, 13.76923076923077, StatsKnn(19.625336940471943, 1.0716500007189238, 2.839082360267639, 9.923076923076923), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), #beam: 4, n: 15
  == Begin Opt. {BeamSearch: bsize=14, hints=0 iteration: 2, expected recall: 0.9, n: 15
  == Iteration finished, current best: (state={BeamSearch: bsize=13, hints=0, score=2.089385474860335, perf=StatsComparison(0.944055944055944, 0.9496503496503497, 0.9467199467199467, 2.6226043701171875e-6, 13.76923076923077, StatsKnn(19.625336940471943, 1.0716500007189238, 2.839082360267639, 9.923076923076923), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), #beam: 4, n: 15
==== END Opt. Finished, best: (state={BeamSearch: bsize=13, hints=0, score=2.089385474860335, perf=StatsComparison(0.944055944055944, 0.9496503496503497, 0.9467199467199467, 2.6226043701171875e-6, 13.76923076923077, StatsKnn(19.625336940471943, 1.0716500007189238, 2.839082360267639, 9.923076923076923), 1.0086939885066106e-6, StatsKnn(19.492548062251164, 1.0716500007189238, 2.7465748511827908, 10.0))  ), n: 15
calling callback optimize_hints; n=31
calling callback optimize_parameters; n=31
==== BEGIN Opt. {BeamSearch: bsize=13, hints=8, expected recall: 0.9, n: 31
  == Begin Opt. {BeamSearch: bsize=13, hints=8 iteration: 1, expected recall: 0.9, n: 31
    ** new best conf: (state={BeamSearch: bsize=11, hints=0, score=2.1193092621664054, perf=StatsComparison(1.0, 1.0, 1.0, 4.395194675611413e-6, 27.695652173913043, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 3, n: 31
    ** new best conf: (state={BeamSearch: bsize=6, hints=0, score=2.1765676567656764, perf=StatsComparison(1.0, 1.0, 1.0, 4.177508146866508e-6, 26.347826086956523, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 4, n: 31
  == Iteration finished, current best: (state={BeamSearch: bsize=6, hints=0, score=2.1765676567656764, perf=StatsComparison(1.0, 1.0, 1.0, 4.177508146866508e-6, 26.347826086956523, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 4, n: 31
  == Begin Opt. {BeamSearch: bsize=9, hints=0 iteration: 2, expected recall: 0.9, n: 31
    ** new best conf: (state={BeamSearch: bsize=4, hints=0, score=2.2619469026548673, perf=StatsComparison(0.9960474308300394, 0.9960474308300394, 0.9960474308300394, 3.3482261326002036e-6, 24.565217391304348, StatsKnn(12.752549669016963, 0.6635502524997877, 1.7087489107380742, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 5, n: 31
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.3817829457364343, perf=StatsComparison(0.992094861660079, 0.992094861660079, 0.992094861660079, 2.912853075110394e-6, 22.434782608695652, StatsKnn(12.779051905092986, 0.6635502524997877, 1.7048270754192187, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 7, n: 31
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.31549815498155, perf=StatsComparison(0.9723320158102765, 0.9723320158102765, 0.9723320158102765, 3.0476114024286685e-6, 23.565217391304348, StatsKnn(12.80672666300898, 0.6635502524997877, 1.7158007258954255, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 8, n: 31
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=2.3817829457364343, perf=StatsComparison(0.992094861660079, 0.992094861660079, 0.992094861660079, 2.912853075110394e-6, 22.434782608695652, StatsKnn(12.779051905092986, 0.6635502524997877, 1.7048270754192187, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 4, n: 31
  == Begin Opt. {BeamSearch: bsize=13, hints=0 iteration: 3, expected recall: 0.9, n: 31
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.4118811881188122, perf=StatsComparison(0.9881422924901185, 0.9881422924901185, 0.9881422924901185, 2.8299248736837637e-6, 21.956521739130434, StatsKnn(12.763230033542799, 0.6635502524997877, 1.7120206252388332, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 5, n: 31
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.3980392156862744, perf=StatsComparison(0.9881422924901185, 0.9881422924901185, 0.9881422924901185, 2.560408219047215e-6, 22.17391304347826, StatsKnn(12.761475024016008, 0.6635502524997877, 1.7132475583449653, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 9, n: 31
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.409090909090909, perf=StatsComparison(0.9644268774703556, 0.9644268774703556, 0.9644268774703556, 2.384185791015625e-6, 22.0, StatsKnn(12.852527079374894, 0.6635502524997877, 1.7369082544160925, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 10, n: 31
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=2.4118811881188122, perf=StatsComparison(0.9881422924901185, 0.9881422924901185, 0.9881422924901185, 2.8299248736837637e-6, 21.956521739130434, StatsKnn(12.763230033542799, 0.6635502524997877, 1.7120206252388332, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), #beam: 4, n: 31
==== END Opt. Finished, best: (state={BeamSearch: bsize=2, hints=0, score=2.4118811881188122, perf=StatsComparison(0.9881422924901185, 0.9881422924901185, 0.9881422924901185, 2.8299248736837637e-6, 21.956521739130434, StatsKnn(12.763230033542799, 0.6635502524997877, 1.7120206252388332, 10.0), 1.482341600501019e-6, StatsKnn(12.742819578751273, 0.6635502524997877, 1.7002965471018916, 10.0))  ), n: 31
calling callback optimize_hints; n=63
calling callback optimize_parameters; n=63
==== BEGIN Opt. {BeamSearch: bsize=2, hints=10, expected recall: 0.9, n: 63
  == Begin Opt. {BeamSearch: bsize=2, hints=10 iteration: 1, expected recall: 0.9, n: 63
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=2.96875, perf=StatsComparison(0.9669421487603306, 0.9669421487603306, 0.9669421487603306, 3.543767062100497e-6, 32.0, StatsKnn(10.25633226741444, 0.5387851396067576, 1.352897817438299, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), #beam: 2, n: 63
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=3.138888888888889, perf=StatsComparison(0.9462809917355371, 0.9462809917355371, 0.9462809917355371, 3.131953152743253e-6, 29.454545454545453, StatsKnn(10.370810161937367, 0.5387851396067576, 1.3903581554239446, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), #beam: 3, n: 63
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=3.0028901734104045, perf=StatsComparison(0.9628099173553721, 0.9628099173553721, 0.9628099173553721, 3.4570693969726562e-6, 31.454545454545453, StatsKnn(10.284456708214499, 0.5387851396067576, 1.3664797977967695, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), #beam: 4, n: 63
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=3.138888888888889, perf=StatsComparison(0.9462809917355371, 0.9462809917355371, 0.9462809917355371, 3.131953152743253e-6, 29.454545454545453, StatsKnn(10.370810161937367, 0.5387851396067576, 1.3903581554239446, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), #beam: 4, n: 63
  == Begin Opt. {BeamSearch: bsize=7, hints=0 iteration: 2, expected recall: 0.9, n: 63
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=3.138888888888889, perf=StatsComparison(0.9462809917355371, 0.9462809917355371, 0.9462809917355371, 3.131953152743253e-6, 29.454545454545453, StatsKnn(10.370810161937367, 0.5387851396067576, 1.3903581554239446, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), #beam: 4, n: 63
==== END Opt. Finished, best: (state={BeamSearch: bsize=2, hints=0, score=3.138888888888889, perf=StatsComparison(0.9462809917355371, 0.9462809917355371, 0.9462809917355371, 3.131953152743253e-6, 29.454545454545453, StatsKnn(10.370810161937367, 0.5387851396067576, 1.3903581554239446, 10.0), 1.9073486328125e-6, StatsKnn(10.198614098808982, 0.5387851396067576, 1.3281498822298916, 10.0))  ), n: 63
calling callback optimize_hints; n=127
calling callback optimize_parameters; n=127
==== BEGIN Opt. {BeamSearch: bsize=2, hints=12, expected recall: 0.9, n: 127
  == Begin Opt. {BeamSearch: bsize=2, hints=12 iteration: 1, expected recall: 0.9, n: 127
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=4.231856738925542, perf=StatsComparison(0.9696969696969696, 0.9696969696969696, 0.9696969696969696, 3.7705456769024884e-6, 39.2962962962963, StatsKnn(8.914100523348209, 0.4990533593076247, 1.1593603513858937, 10.0), 2.887513902452257e-6, StatsKnn(8.873191709871646, 0.4990533593076247, 1.1478319388848763, 10.0))  ), #beam: 4, n: 127
==== END Opt. Finished, best: (state={BeamSearch: bsize=2, hints=0, score=4.231856738925542, perf=StatsComparison(0.9696969696969696, 0.9696969696969696, 0.9696969696969696, 3.7705456769024884e-6, 39.2962962962963, StatsKnn(8.914100523348209, 0.4990533593076247, 1.1593603513858937, 10.0), 2.887513902452257e-6, StatsKnn(8.873191709871646, 0.4990533593076247, 1.1478319388848763, 10.0))  ), n: 127
calling callback optimize_hints; n=255
calling callback optimize_parameters; n=255
==== BEGIN Opt. {BeamSearch: bsize=2, hints=13, expected recall: 0.9, n: 255
  == Begin Opt. {BeamSearch: bsize=2, hints=13 iteration: 1, expected recall: 0.9, n: 255
  == Iteration finished, current best: (state={BeamSearch: bsize=10, hints=0, score=5.981096408317581, perf=StatsComparison(0.9736070381231671, 0.9736070381231671, 0.9736070381231671, 5.445172709803428e-6, 51.193548387096776, StatsKnn(6.144473883413499, 0.3335683422703897, 0.787967509800388, 10.0), 4.353061799080141e-6, StatsKnn(6.124145438594203, 0.3335683422703897, 0.7794910121348596, 10.0))  ), #beam: 4, n: 255
==== END Opt. Finished, best: (state={BeamSearch: bsize=10, hints=0, score=5.981096408317581, perf=StatsComparison(0.9736070381231671, 0.9736070381231671, 0.9736070381231671, 5.445172709803428e-6, 51.193548387096776, StatsKnn(6.144473883413499, 0.3335683422703897, 0.787967509800388, 10.0), 4.353061799080141e-6, StatsKnn(6.124145438594203, 0.3335683422703897, 0.7794910121348596, 10.0))  ), n: 255
calling callback optimize_hints; n=511
calling callback optimize_parameters; n=511
==== BEGIN Opt. {BeamSearch: bsize=10, hints=16, expected recall: 0.9, n: 511
  == Begin Opt. {BeamSearch: bsize=10, hints=16 iteration: 1, expected recall: 0.9, n: 511
    ** new best conf: (state={BeamSearch: bsize=5, hints=0, score=8.490609253321118, perf=StatsComparison(1.0, 1.0, 1.0, 7.68899917602539e-6, 68.21875, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 3, n: 511
    ** new best conf: (state={BeamSearch: bsize=5, hints=0, score=8.480329368709972, perf=StatsComparison(1.0, 1.0, 1.0, 7.309019565582275e-6, 68.3125, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 4, n: 511
    ** new best conf: (state={BeamSearch: bsize=4, hints=0, score=8.884281581485052, perf=StatsComparison(0.9943181818181818, 0.9943181818181818, 0.9943181818181818, 6.191432476043701e-6, 64.8125, StatsKnn(6.099544197320938, 0.3432123304810375, 0.7710073888301849, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 5, n: 511
  == Iteration finished, current best: (state={BeamSearch: bsize=4, hints=0, score=8.884281581485052, perf=StatsComparison(0.9943181818181818, 0.9943181818181818, 0.9943181818181818, 6.191432476043701e-6, 64.8125, StatsKnn(6.099544197320938, 0.3432123304810375, 0.7710073888301849, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 4, n: 511
  == Begin Opt. {BeamSearch: bsize=17, hints=0 iteration: 2, expected recall: 0.9, n: 511
    ** new best conf: (state={BeamSearch: bsize=3, hints=0, score=8.949440933398153, perf=StatsComparison(0.9914772727272728, 0.9914772727272728, 0.9914772727272728, 6.623566150665283e-6, 64.28125, StatsKnn(6.099450007081032, 0.3432123304810375, 0.7717570532113314, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 8, n: 511
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=8.95330739299611, perf=StatsComparison(0.9886363636363636, 0.9886363636363636, 0.9886363636363636, 5.535781383514404e-6, 64.25, StatsKnn(6.101091586053371, 0.3432123304810375, 0.7728991974145174, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 10, n: 511
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=9.131277971158628, perf=StatsComparison(0.9886363636363636, 0.9886363636363636, 0.9886363636363636, 6.370246410369873e-6, 62.84375, StatsKnn(6.100984498858452, 0.3432123304810375, 0.772439805790782, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 21, n: 511
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=9.131277971158628, perf=StatsComparison(0.9886363636363636, 0.9886363636363636, 0.9886363636363636, 6.370246410369873e-6, 62.84375, StatsKnn(6.100984498858452, 0.3432123304810375, 0.772439805790782, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 4, n: 511
  == Begin Opt. {BeamSearch: bsize=13, hints=0 iteration: 3, expected recall: 0.9, n: 511
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=9.159680638722556, perf=StatsComparison(0.9829545454545455, 0.9829545454545455, 0.9829545454545455, 6.310641765594482e-6, 62.625, StatsKnn(6.107482016086578, 0.3432123304810375, 0.773013036698103, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 5, n: 511
  == Iteration finished, current best: (state={BeamSearch: bsize=2, hints=0, score=9.159680638722556, perf=StatsComparison(0.9829545454545455, 0.9829545454545455, 0.9829545454545455, 6.310641765594482e-6, 62.625, StatsKnn(6.107482016086578, 0.3432123304810375, 0.773013036698103, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), #beam: 4, n: 511
==== END Opt. Finished, best: (state={BeamSearch: bsize=2, hints=0, score=9.159680638722556, perf=StatsComparison(0.9829545454545455, 0.9829545454545455, 0.9829545454545455, 6.310641765594482e-6, 62.625, StatsKnn(6.107482016086578, 0.3432123304810375, 0.773013036698103, 10.0), 6.966292858123779e-6, StatsKnn(6.093612626194954, 0.3432123304810375, 0.7699197512120008, 10.0))  ), n: 511
calling callback optimize_hints; n=1023
calling callback optimize_parameters; n=1023
==== BEGIN Opt. {BeamSearch: bsize=2, hints=18, expected recall: 0.9, n: 1023
  == Begin Opt. {BeamSearch: bsize=2, hints=18 iteration: 1, expected recall: 0.9, n: 1023
  == Iteration finished, current best: (state={BeamSearch: bsize=10, hints=0, score=16.180947821924363, perf=StatsComparison(0.9824046920821116, 0.9824046920821116, 0.9824046920821116, 6.321937807144657e-6, 67.38709677419355, StatsKnn(4.726883826717254, 0.266318135804707, 0.6047846299986686, 10.0), 1.1321037046370968e-5, StatsKnn(4.716615353861163, 0.266318135804707, 0.6015012898752766, 10.0))  ), #beam: 4, n: 1023
==== END Opt. Finished, best: (state={BeamSearch: bsize=10, hints=0, score=16.180947821924363, perf=StatsComparison(0.9824046920821116, 0.9824046920821116, 0.9824046920821116, 6.321937807144657e-6, 67.38709677419355, StatsKnn(4.726883826717254, 0.266318135804707, 0.6047846299986686, 10.0), 1.1321037046370968e-5, StatsKnn(4.716615353861163, 0.266318135804707, 0.6015012898752766, 10.0))  ), n: 1023
calling callback optimize_hints; n=2047
calling callback optimize_parameters; n=2047
==== BEGIN Opt. {BeamSearch: bsize=10, hints=19, expected recall: 0.9, n: 2047
  == Begin Opt. {BeamSearch: bsize=10, hints=19 iteration: 1, expected recall: 0.9, n: 2047
    ** new best conf: (state={BeamSearch: bsize=7, hints=0, score=20.30562923666372, perf=StatsComparison(1.0, 1.0, 1.0, 1.1496245861053467e-5, 106.03125, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 2, n: 2047
    ** new best conf: (state={BeamSearch: bsize=8, hints=0, score=19.312552418227565, perf=StatsComparison(1.0, 1.0, 1.0, 1.3127923011779785e-5, 111.78125, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 6, n: 2047
    ** new best conf: (state={BeamSearch: bsize=10, hints=0, score=20.036326649229874, perf=StatsComparison(1.0, 1.0, 1.0, 1.274794340133667e-5, 107.53125, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 7, n: 2047
    ** new best conf: (state={BeamSearch: bsize=9, hints=0, score=19.462232243517473, perf=StatsComparison(1.0, 1.0, 1.0, 1.268833875656128e-5, 110.875, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 8, n: 2047
    ** new best conf: (state={BeamSearch: bsize=3, hints=0, score=23.02555480833894, perf=StatsComparison(0.9971590909090909, 0.9971590909090909, 0.9971590909090909, 9.037554264068604e-6, 92.9375, StatsKnn(3.709390088915825, 0.21645535889547318, 0.46813064627349377, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 9, n: 2047
    ** new best conf: (state={BeamSearch: bsize=3, hints=0, score=20.617849655585506, perf=StatsComparison(0.9971590909090909, 0.9971590909090909, 0.9971590909090909, 9.745359420776367e-6, 104.34375, StatsKnn(3.710079811513424, 0.21645535889547318, 0.4683134639635682, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 10, n: 2047
    ** new best conf: (state={BeamSearch: bsize=5, hints=0, score=19.281886687133685, perf=StatsComparison(1.0, 1.0, 1.0, 1.0117888450622559e-5, 111.96875, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 11, n: 2047
  == Iteration finished, current best: (state={BeamSearch: bsize=3, hints=0, score=23.02555480833894, perf=StatsComparison(0.9971590909090909, 0.9971590909090909, 0.9971590909090909, 9.037554264068604e-6, 92.9375, StatsKnn(3.709390088915825, 0.21645535889547318, 0.46813064627349377, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 4, n: 2047
  == Begin Opt. {BeamSearch: bsize=2, hints=0 iteration: 2, expected recall: 0.9, n: 2047
  == Iteration finished, current best: (state={BeamSearch: bsize=3, hints=0, score=23.02555480833894, perf=StatsComparison(0.9971590909090909, 0.9971590909090909, 0.9971590909090909, 9.037554264068604e-6, 92.9375, StatsKnn(3.709390088915825, 0.21645535889547318, 0.46813064627349377, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), #beam: 4, n: 2047
==== END Opt. Finished, best: (state={BeamSearch: bsize=3, hints=0, score=23.02555480833894, perf=StatsComparison(0.9971590909090909, 0.9971590909090909, 0.9971590909090909, 9.037554264068604e-6, 92.9375, StatsKnn(3.709390088915825, 0.21645535889547318, 0.46813064627349377, 10.0), 1.9937753677368164e-5, StatsKnn(3.7093411460518837, 0.21645535889547318, 0.4680817034095526, 10.0))  ), n: 2047
calling callback optimize_hints; n=4095
calling callback optimize_parameters; n=4095
==== BEGIN Opt. {BeamSearch: bsize=3, hints=21, expected recall: 0.9, n: 4095
  == Begin Opt. {BeamSearch: bsize=3, hints=21 iteration: 1, expected recall: 0.9, n: 4095
  == Iteration finished, current best: (state={BeamSearch: bsize=16, hints=0, score=43.78564206268959, perf=StatsComparison(1.0, 1.0, 1.0, 1.0482726558562248e-5, 95.70967741935483, StatsKnn(3.2399219197611653, 0.16624526042611368, 0.41568919391401354, 10.0), 4.122334141885081e-5, StatsKnn(3.2399219197611653, 0.16624526042611368, 0.41568919391401354, 10.0))  ), #beam: 4, n: 4095
==== END Opt. Finished, best: (state={BeamSearch: bsize=16, hints=0, score=43.78564206268959, perf=StatsComparison(1.0, 1.0, 1.0, 1.0482726558562248e-5, 95.70967741935483, StatsKnn(3.2399219197611653, 0.16624526042611368, 0.41568919391401354, 10.0), 4.122334141885081e-5, StatsKnn(3.2399219197611653, 0.16624526042611368, 0.41568919391401354, 10.0))  ), n: 4095
calling callback optimize_hints; n=8191
calling callback optimize_parameters; n=8191
==== BEGIN Opt. {BeamSearch: bsize=16, hints=23, expected recall: 0.9, n: 8191
  == Begin Opt. {BeamSearch: bsize=16, hints=23 iteration: 1, expected recall: 0.9, n: 8191
    ** new best conf: (state={BeamSearch: bsize=5, hints=0, score=56.816013628620105, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.7561018466949463e-5, 146.75, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 2, n: 8191
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=48.43250090481361, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.8998980522155762e-5, 172.6875, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 3, n: 8191
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=47.99874484489869, perf=StatsComparison(0.9943181818181819, 0.9943181818181819, 0.9943181818181819, 1.8440186977386475e-5, 174.28125, StatsKnn(2.315124362707138, 0.13207409402821213, 0.29538334161043167, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 4, n: 8191
    ** new best conf: (state={BeamSearch: bsize=6, hints=0, score=61.75846082522021, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.6532838344573975e-5, 134.8125, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 5, n: 8191
    ** new best conf: (state={BeamSearch: bsize=13, hints=0, score=48.6653937079469, perf=StatsComparison(1.0, 1.0, 1.0, 2.0444393157958984e-5, 171.84375, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 6, n: 8191
    ** new best conf: (state={BeamSearch: bsize=4, hints=0, score=52.505600314403615, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.5847384929656982e-5, 159.03125, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 7, n: 8191
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=48.23589836006488, perf=StatsComparison(0.9943181818181819, 0.9943181818181819, 0.9943181818181819, 1.703202724456787e-5, 173.40625, StatsKnn(2.315286912024021, 0.13207409402821213, 0.29502179473638535, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 8, n: 8191
    ** new best conf: (state={BeamSearch: bsize=2, hints=0, score=48.167896346949796, perf=StatsComparison(0.9943181818181819, 0.9943181818181819, 0.9943181818181819, 1.7344951629638672e-5, 173.65625, StatsKnn(2.318550854921341, 0.13207409402821213, 0.29538334161043167, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 9, n: 8191
    ** new best conf: (state={BeamSearch: bsize=12, hints=0, score=52.55625491738788, perf=StatsComparison(1.0, 1.0, 1.0, 2.0347535610198975e-5, 158.875, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 10, n: 8191
  == Iteration finished, current best: (state={BeamSearch: bsize=6, hints=0, score=61.75846082522021, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.6532838344573975e-5, 134.8125, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 4, n: 8191
  == Begin Opt. {BeamSearch: bsize=12, hints=0 iteration: 2, expected recall: 0.9, n: 8191
  == Iteration finished, current best: (state={BeamSearch: bsize=6, hints=0, score=61.75846082522021, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.6532838344573975e-5, 134.8125, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), #beam: 4, n: 8191
==== END Opt. Finished, best: (state={BeamSearch: bsize=6, hints=0, score=61.75846082522021, perf=StatsComparison(0.9971590909090908, 0.9971590909090908, 0.9971590909090908, 1.6532838344573975e-5, 134.8125, StatsKnn(2.3105986192822456, 0.13207409402821213, 0.2937600240111351, 10.0), 8.406490087509155e-5, StatsKnn(2.3105141520500183, 0.13207409402821213, 0.2936755567789078, 10.0))  ), n: 8191
added n=10000, neighborhood=14, {BeamSearch: bsize=6, hints=0, LogNeighborhood, 2021-07-01T00:05:35.765
KnnResult{Int32, Float32}(Int32[4272, 5664, 4742], Float32[0.46436483, 0.6274073, 0.8162391], 3, 0)
KnnResult{Int32, Float32}(Int32[1188, 9480, 4114], Float32[0.024345834, 0.14145628, 0.14444777], 3, 0)
KnnResult{Int32, Float32}(Int32[4845, 2037, 1175], Float32[0.07080072, 0.07350176, 0.1246301], 3, 0)
KnnResult{Int32, Float32}(Int32[9421, 7719, 5659], Float32[0.12732568, 0.1484435, 0.22651052], 3, 0)
KnnResult{Int32, Float32}(Int32[39, 1047, 7777], Float32[0.07404027, 0.08341181, 0.10041868], 3, 0)

There are a several options for the construction of the index, from specifying the precise searching algorithm (search_algo) and the neighborhood strategy to be used on construction (neighborhood_algo). You can also specify here if the construction will be made in parallel and how this paralllelism will be performed (firstblock, block).

Missing docstring.

Missing docstring for SearchGraph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SearchGraphOptions. Check Documenter's build log for details.

Local search algorithms

Regarding the search search algorithm, there are two alternatives, based on local search heuristics.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, hints=Int32[], beam=KnnResult(bsize), vstate=VisitedVertices())

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time. Multithreading applications must have copies of this object due to shared cache objects.

hints: An initial hint for the exploration (empty hints imply bsize random starting points).

bsize: The size of the beam.

beam: A cache object for reducing memory allocations

vstate: A cache object for reducing memory allocations

SimilaritySearch.IHCSearch — Type

IHCSearch(hints::Vector; restarts=length(hints), localimprovements=false)
IHCSearch(restarts::Integer=20; hints=Int32[], localimprovements=false)
IHCSearch(hints, restarts, localimprovements, vstate)

IHCSearch is an iterated hill climbing algorithma, a local search algorithm. It greedily navigates the search graph and restart the search restarts times. Multithreading applications must have copies of this object due to shared cache objects.

restarts: The number of restarts.

hints: An initial hint for the exploration (if it is not empty, then superseeds the initial points of the random starting points).

localimprovements: An experimental technique that if it is true it will achieve very high quality results, at cost of increasing searching time.

vstate: A cache object for reducing memory allocations

Neighborhood algorithms

Regarding neighborhood algorithms, the package defines the following ones:

Neighborhood algorithms

SimilaritySearch.FixedNeighborhood — Type

FixedNeighborhood(k=8)

New items are connected with a fixed number of neighbors

SimilaritySearch.LogNeighborhood — Type

LogNeighborhood(b=2)

New items are connected with the $\log_b n$ items in the current graph (where $n$ is the current number of vertices in the graph)

SimilaritySearch.LogSatNeighborhood — Type

LogSatNeighborhood(b=1.1)

New items are connected with a half partitioning scheme based on the Spatial access tree (cite), $\log_b n$ near items are considered for this procedure. The Log-Sat method will produce small neighborhood with a nice covering of the search space.

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood(k=32)

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

Incremental construction of the index

The index supports insertions; for instance, it can be fully constructed using insertions as follows:

using SimilaritySearch

index = SearchGraph()

append!(index, [rand(Float32, 4) for i in 1:1000])

search(index, rand(Float32, 4), 3)

KnnResult{Int32, Float32}(Int32[718, 313, 358], Float32[0.0064238105, 0.012094451, 0.024711695], 3, 0)

Each insertion is performed by the push! function or a chunk of items with append!; the last function also allows parallel insertions.

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

Base.append! — Method

append!(index::SearchGraph, db; parallel=false, parallel_firstblock=30_000, parallel_block=10_000)

Appends all items in db to the index. It can be made in parallel or sequentially. In case of a parallel appending, then parallel_firstblock indicates the minimum number of items before going parallel, and parallel_block sets the chunck size to append in parallel.

Note: Parallel construction doesn't trigger callbacks listed in `callback_list', they must be executed manually.

Customizing the insertion method

In particular, the procedure of pushing items into the index is made calling find_neighborhood and push_neighborhood; these methods can be overriden or used in other wyas to customize the insertion of data and the construction of the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(algo::FixedNeighborhood, index::SearchGraph, item)

Finds a list of neighbors using the FixedNeighborhood criterion of item in the index

find_neighborhood(index::SearchGraph{T}, item)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, neighbors::AbstractVector{Int32}; apply_callbacks=true)

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

Optimizing the index's performance

SimilaritySearch.optimize! — Function

function optimize!(search_algo::LocalSearchAlgorithm,
                   index::SearchGraph{T},
                   recall::Float64,
                   perf::Performance;
                   bsize::Int=4,
                   tol::Float64=0.01,
                   probes::Int=0) where T

Optimizes a local search index for an specific algorithm to get the desired performance. Note that optimizing for low-recall will yield to faster searches; the train queries are specified as part of the perf struct.