SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(L2Distance(), X, verbose=true)

for q in Q
    println(search(index, q, KnnResult(3)))
end

{SearchGraph: dist=L2Distance(), n=0, search_algo={BeamSearch: bsize=16, ssize=16, hints=0, neighborhood_algo=LogNeighborhood, knn=10}
added n=10000, neighborhood=14, {BeamSearch: bsize=16, ssize=16, hints=0, LogNeighborhood, 2021-02-23T19:58:33.276
KnnResult(3, 3, Item[Item(402, 0.10998304f0), Item(5532, 0.18648586f0), Item(5117, 0.33306184f0)])
KnnResult(3, 3, Item[Item(9878, 0.102121234f0), Item(4410, 0.10767361f0), Item(748, 0.13991475f0)])
KnnResult(3, 3, Item[Item(5844, 0.071972564f0), Item(8910, 0.0859385f0), Item(429, 0.10994555f0)])
KnnResult(3, 3, Item[Item(6635, 0.22495024f0), Item(7647, 0.25874647f0), Item(5981, 0.26647466f0)])
KnnResult(3, 3, Item[Item(1156, 0.107208155f0), Item(7481, 0.12451083f0), Item(7635, 0.14951484f0)])

There are a several options for the construction of the index.

SimilaritySearch.SearchGraph — Type

SearchGraph(dist::PreMetric,
    db::AbstractVector;
    search_algo::LocalSearchAlgorithm=BeamSearch(),
    neighborhood_algo::NeighborhoodAlgorithm=LogNeighborhood(),
    automatic_optimization=false,
    recall=0.9,
    ksearch=10,
    tol=0.001,
    verbose=true)

Creates a SearchGraph object, i.e., an index to perform approximate search on db using the given search and neighbohood strategies. If automatic_optimization is true, then the structure tries to reach the given recall under the given ksearch.

SimilaritySearch.SearchGraphOptions — Type

SearchGraphOptions(automatic_optimization::Bool, recall::Float64, ksearch::Int, tol::Float64, verbose::Bool)

Defines a number of options for the SearchGraph

Please note

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, ssize=bsize; hints=Int32[], beam=KnnResult(bsize), vstate=VisitedVertices())

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time. Multithreading applications must have copies of this object due to shared cache objects.

ssize: The size of the first sampling.

hints: An initial hint for the exploration (if it is not empty, then superseeds ssize and the initial sampling).

beam: A cache object for reducing memory allocations

vstate: A cache object for reducing memory allocations

SimilaritySearch.IHCSearch — Type

IHCSearch(hints::Vector; restarts=length(hints), localimprovements=false)
IHCSearch(restarts::Integer=20; hints=Int32[], localimprovements=false)
IHCSearch(hints, restarts, localimprovements, vstate)

IHCSearch is an iterated hill climbing algorithma, a local search algorithm. It greedily navigates the search graph and restart the search restarts times. Multithreading applications must have copies of this object due to shared cache objects.

restarts: The number of restarts.

hints: An initial hint for the exploration (if it is not empty, then superseeds the initial points of the random starting points).

localimprovements: An experimental technique that if it is true it will achieve very high quality results, at cost of increasing searching time.

vstate: A cache object for reducing memory allocations

Neighborhood algorithms

The index can be created incrementally,

using SimilaritySearch

index = SearchGraph(L2Distance(), Vector{Float32}[])

for i in 1:1000
    push!(index, rand(Float32, 4))
end

search(index, rand(Float32, 4), 3)

KnnResult(3, 3, Item[Item(564, 0.13394675f0), Item(803, 0.16389982f0), Item(867, 0.16730326f0), Item(686, 0.20802914f0), Item(969, 0.20871653f0), Item(302, 0.21621701f0), Item(640, 0.21675986f0), Item(51, 0.21949938f0), Item(588, 0.22114432f0), Item(472, 0.23312618f0)])

The index construction can be customized with find_neighborhood and push_neighborhood

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(algo::FixedNeighborhood, index::SearchGraph, item)

Finds a list of neighbors using the FixedNeighborhood criterion of item in the index

find_neighborhood(index::SearchGraph{T}, item)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, L::AbstractVector{Int32})

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

SimilaritySearch.optimize! — Function

function optimize!(search_algo::LocalSearchAlgorithm,
                   index::SearchGraph{T},
                   recall::Float64,
                   perf::Performance;
                   bsize::Int=4,
                   tol::Float64=0.01,
                   probes::Int=0) where T

Optimizes a local search index for an specific algorithm to get the desired performance. Note that optimizing for low-recall will yield to faster searches; the train queries are specified as part of the perf struct.

also see search and push!