SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(; dist=L2Distance(), verbose=true)
append!(index, X)

for q in Q
    println(search(index, q, KnnResult(3)))
end

calling callback hints; n=8, type=RandomHintsCallback
calling callback neighborhood; n=8, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=16, type=RandomHintsCallback
calling callback neighborhood; n=16, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=32, type=RandomHintsCallback
calling callback neighborhood; n=32, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=64, type=RandomHintsCallback
calling callback neighborhood; n=64, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=128, type=RandomHintsCallback
calling callback neighborhood; n=128, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=256, type=RandomHintsCallback
calling callback neighborhood; n=256, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=512, type=RandomHintsCallback
calling callback neighborhood; n=512, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=1024, type=RandomHintsCallback
calling callback neighborhood; n=1024, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=2048, type=RandomHintsCallback
calling callback neighborhood; n=2048, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=4096, type=RandomHintsCallback
calling callback neighborhood; n=4096, type=SimilaritySearch.NeighborhoodCallback
calling callback hints; n=8192, type=RandomHintsCallback
calling callback neighborhood; n=8192, type=SimilaritySearch.NeighborhoodCallback
Pair{Int32, Float32}[916 => 0.3907658, 1673 => 0.46853542, 2034 => 0.51899934]
Pair{Int32, Float32}[1226 => 0.08222949, 6651 => 0.5527809, 1458 => 0.60658306]
Pair{Int32, Float32}[2104 => 0.43459567, 8252 => 0.65733206, 9671 => 0.96223056]
Pair{Int32, Float32}[172 => 0.18255138, 278 => 0.20035684, 1268 => 0.34314018]
Pair{Int32, Float32}[421 => 0.75640893, 778 => 0.84713715, 65 => 0.88513947]

There are a several options for the construction of the index, from specifying the precise searching algorithm (search_algo) and the neighborhood strategy to be used on construction (neighborhood_algo). You can also specify here if the construction will be made in parallel and how this paralllelism will be performed (firstblock, block).

SimilaritySearch.SearchGraph — Type

struct SearchGraph <: AbstractSearchContext

SearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.

Note 1: Parallel searches need copies of the structure or passing a KnnResult per thread. Note 2: Parallel insertions should be made through append! function with parallel=true

Missing docstring.

Missing docstring for SearchGraphOptions. Check Documenter's build log for details.

Local search algorithms

Regarding the search search algorithm, there are two alternatives, based on local search heuristics.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, hints=Int32[], beam=KnnResult(bsize), vstate=VisitedVertices())

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time. Multithreading applications must have copies of this object due to shared cache objects.

hints: An initial hint for the exploration (empty hints imply bsize random starting points).

bsize: The size of the beam.

beam: A cache object for reducing memory allocations

vstate: A cache object for reducing memory allocations

SimilaritySearch.IHCSearch — Type

IHCSearch(hints::Vector; restarts=length(hints), localimprovements=false)
IHCSearch(restarts::Integer=20; hints=Int32[], localimprovements=false)
IHCSearch(hints, restarts, localimprovements, vstate)

IHCSearch is an iterated hill climbing algorithma, a local search algorithm. It greedily navigates the search graph and restart the search restarts times. Multithreading applications must have copies of this object due to shared cache objects.

restarts: The number of restarts.

hints: An initial hint for the exploration (if it is not empty, then superseeds the initial points of the random starting points).

localimprovements: An experimental technique that if it is true it will achieve very high quality results, at cost of increasing searching time.

vstate: A cache object for reducing memory allocations

Neighborhood algorithms

Regarding neighborhood algorithms, the package defines the following ones:

Neighborhood algorithms

Missing docstring.

Missing docstring for FixedNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogSatNeighborhood. Check Documenter's build log for details.

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood(k=32)

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

Incremental construction of the index

The index supports insertions; for instance, it can be fully constructed using insertions as follows:

using SimilaritySearch

index = SearchGraph()

append!(index, [rand(Float32, 4) for i in 1:1000])

search(index, rand(Float32, 4), 3)

3-element KnnResult{Int32, Float32}:
 49 => 0.024652382
 74 => 0.031942226
 87 => 0.039689317

Each insertion is performed by the push! function or a chunk of items with append!; the last function also allows parallel insertions.

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

Base.append! — Method

append!(index::SearchGraph, db; parallel_block=1, parallel_firstblock=parallel_block, apply_callbacks=true)

Appends all items in db to the index. It can be made in parallel or sequentially. In case of a parallel appending, then:

parallel_block must be bigger than 1 and describes the batch size to append in parallel (i.e., in the order of thousands, depending on the size of the db and the number of available threads).

parallel_firstblock indicates the minimum number of items inserted sequentially before going parallel, it can be 0 if the index is already populated, defaults to parallel_block.

Note: Parallel doesn't trigger callbacks inside blocks.

Customizing the insertion method

In particular, the procedure of pushing items into the index is made calling find_neighborhood and push_neighborhood; these methods can be overriden or used in other wyas to customize the insertion of data and the construction of the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(index::SearchGraph{T}, item; res=index.res)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, neighbors::KnnResult; apply_callbacks=true)

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

Optimizing the index's performance

Missing docstring.

Missing docstring for optimize!. Check Documenter's build log for details.