SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(; dist=L2Distance(), verbose=true)
append!(index, X)

for q in Q
    println(search(index, q, KnnResult(3)))
end

calling callback DisjointHints at n=11
calling callback NeighborhoodSize at n=11
calling callback DisjointHints at n=17
calling callback NeighborhoodSize at n=17
calling callback DisjointHints at n=25
calling callback NeighborhoodSize at n=25
calling callback DisjointHints at n=38
calling callback NeighborhoodSize at n=38
calling callback DisjointHints at n=57
calling callback NeighborhoodSize at n=57
calling callback DisjointHints at n=86
calling callback NeighborhoodSize at n=86
calling callback DisjointHints at n=129
calling callback NeighborhoodSize at n=129
calling callback DisjointHints at n=194
calling callback NeighborhoodSize at n=194
calling callback DisjointHints at n=291
calling callback NeighborhoodSize at n=291
calling callback DisjointHints at n=437
calling callback NeighborhoodSize at n=437
calling callback DisjointHints at n=656
calling callback NeighborhoodSize at n=656
calling callback DisjointHints at n=985
calling callback NeighborhoodSize at n=985
calling callback DisjointHints at n=1477
calling callback NeighborhoodSize at n=1477
calling callback DisjointHints at n=2216
calling callback NeighborhoodSize at n=2216
calling callback DisjointHints at n=3325
calling callback NeighborhoodSize at n=3325
calling callback DisjointHints at n=4987
calling callback NeighborhoodSize at n=4987
calling callback DisjointHints at n=7481
calling callback NeighborhoodSize at n=7481
(res = KnnResult(Int32[9565, 8579, 2558], Float32[0.09848699, 0.159677, 0.16358984], 3), cost = 214)
(res = KnnResult(Int32[442, 8128, 305], Float32[0.037119858, 0.10394967, 0.1074205], 3), cost = 231)
(res = KnnResult(Int32[3591, 5478, 3478], Float32[0.04747759, 0.08438669, 0.11445825], 3), cost = 237)
(res = KnnResult(Int32[9747, 6816, 1108], Float32[0.071226254, 0.10168051, 0.14134367], 3), cost = 188)
(res = KnnResult(Int32[2381, 9519, 565], Float32[0.04062478, 0.07003241, 0.086439826], 3), cost = 222)

There are a several options for the construction of the index, from specifying the precise searching algorithm (search_algo) and the neighborhood strategy to be used on construction (neighborhood_algo). You can also specify here if the construction will be made in parallel and how this paralllelism will be performed (firstblock, block).

SimilaritySearch.SearchGraph — Type

struct SearchGraph <: AbstractSearchContext

SearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.

hints: Initial points for exploration (empty hints imply using random points)

Note: Parallel insertions should be made through append! or index! function with parallel_block > 1

Missing docstring.

Missing docstring for SearchGraphOptions. Check Documenter's build log for details.

Local search algorithms

Regarding the search search algorithm, there are two alternatives, based on local search heuristics.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, Δ::Float32)

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time.

bsize: The size of the beam.

Δ: Soft margin for accepting elements into the beam

maxvisits: MAximum visits while searching, useful for early stopping without convergence

Neighborhood algorithms

Regarding neighborhood algorithms, the package defines the following ones:

Neighborhood algorithms

Missing docstring.

Missing docstring for FixedNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogSatNeighborhood. Check Documenter's build log for details.

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood()

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

Incremental construction of the index

The index supports insertions; for instance, it can be fully constructed using insertions as follows:

using SimilaritySearch

index = SearchGraph()

append!(index, [rand(Float32, 4) for i in 1:1000])

search(index, rand(Float32, 4), 3)

Each insertion is performed by the push! function or a chunk of items with append!; the last function also allows parallel insertions.

Base.push! — Method

push!(index::SearchGraph, item; push_item=true, pools=getpools(index))

Appends item into the index.

Base.append! — Method

append!(index::SearchGraph, db; parallel_block=1, parallel_minimum_first_block=parallel_block, apply_callbacks=true)

Appends all items in db to the index. It can be made in parallel or sequentially. In case of a parallel appending, then:

parallel_block must be bigger than 1 and describes the batch size to append in parallel (i.e., in the order of thousands, depending on the size of the db and the number of available threads).

parallel_minimum_first_block indicates the minimum number of items inserted sequentially before going parallel, it can be 0 if the index is already populated, defaults to parallel_block.

Note: Parallel doesn't trigger callbacks inside blocks.

Customizing the insertion method

In particular, the procedure of pushing items into the index is made calling find_neighborhood and push_neighborhood; these methods can be overriden or used in other wyas to customize the insertion of data and the construction of the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(index::SearchGraph{T}, item, pools; hints=index.hints)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function). res is always reused since reduce creates a new KnnResult from it (a copy if reduce in its simpler terms)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, neighbors; push_item=true, apply_callbacks=true)

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

Optimizing the index's performance

SimilaritySearch.optimize! — Function

optimize!(
    index::SearchGraph,
    opt::OptimizeParameters;
    queries=nothing,
    scalemaxvisits=1.0,
    maxvisits=2 * index.search_algo.maxvisits,
    verbose=index.verbose
)

Optimizes the index using the opt parameters. If queries=nothing then it selects a small sample randomply from the already indexed objects; this sample size and the number of neighbors are also parameters in opt.

Note: The visits parameters is a factor allowing early stopping based on the number of distance evaluations seen in the optimization procedure; please note that in some cases this will reduce the quality to limit the search times. Please also take into account that inserting items after limiting visits may also cause severe quality degradation when maxvisits is not also updated as required. You can always adjust maxvisits modifying index.search_algo.maxvisits.