SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(L2Distance(), X, verbose=true)

for q in Q
    println(search(index, q, KnnResult(3)))
end

{SearchGraph: dist=L2Distance(), n=0, search_algo={BeamSearch: bsize=16, ssize=16, hints=0, neighborhood_algo=LogNeighborhood, knn=10}
added n=10000, neighborhood=14, {BeamSearch: bsize=16, ssize=16, hints=0, LogNeighborhood, 2021-04-08T16:50:15.070
KnnResult(3, 3, Item[Item(9601, 0.39779282f0), Item(5211, 0.42952293f0), Item(9424, 0.47298467f0)])
KnnResult(3, 3, Item[Item(7451, 0.09358546f0), Item(9155, 0.11530786f0), Item(8325, 0.116146125f0)])
KnnResult(3, 3, Item[Item(9946, 0.18577775f0), Item(8105, 0.2749636f0), Item(1592, 0.2955328f0)])
KnnResult(3, 3, Item[Item(8674, 0.21437457f0), Item(5171, 0.2755346f0), Item(5364, 0.30289927f0)])
KnnResult(3, 3, Item[Item(6640, 0.07491649f0), Item(2371, 0.19897696f0), Item(7297, 0.20428735f0)])

There are a several options for the construction of the index, from specifying the precise searching algorithm (search_algo) and the neighborhood strategy to be used on construction (neighborhood_algo). You can also specify here if the construction will be made in parallel and how this paralllelism will be performed (firstblock, block).

SimilaritySearch.SearchGraph — Type

SearchGraph(dist::PreMetric,
    db::AbstractVector;
    search_algo::LocalSearchAlgorithm=BeamSearch(),
    neighborhood_algo::NeighborhoodAlgorithm=LogNeighborhood(),
    automatic_optimization=false,
    recall=0.9,
    ksearch=10,
    tol=0.001,
    parallel=false,
    firstblock=100_000,
    block=10_000, 
    verbose=true)

Creates a SearchGraph object, i.e., an index to perform approximate search on db using the given search and neighbohood strategies. If automatic_optimization is true, then the structure tries to reach the given recall under the given ksearch. The construction will use all available threads if parallel=true.

SimilaritySearch.SearchGraphOptions — Type

SearchGraphOptions(automatic_optimization::Bool, recall::Float64, ksearch::Int, tol::Float64, verbose::Bool)

Defines a number of options for the SearchGraph

Local search algorithms

Regarding the search search algorithm, there are two alternatives, based on local search heuristics.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, ssize=bsize; hints=Int32[], beam=KnnResult(bsize), vstate=VisitedVertices())

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time. Multithreading applications must have copies of this object due to shared cache objects.

ssize: The size of the first sampling.

hints: An initial hint for the exploration (if it is not empty, then superseeds ssize and the initial sampling).

beam: A cache object for reducing memory allocations

vstate: A cache object for reducing memory allocations

SimilaritySearch.IHCSearch — Type

IHCSearch(hints::Vector; restarts=length(hints), localimprovements=false)
IHCSearch(restarts::Integer=20; hints=Int32[], localimprovements=false)
IHCSearch(hints, restarts, localimprovements, vstate)

IHCSearch is an iterated hill climbing algorithma, a local search algorithm. It greedily navigates the search graph and restart the search restarts times. Multithreading applications must have copies of this object due to shared cache objects.

restarts: The number of restarts.

hints: An initial hint for the exploration (if it is not empty, then superseeds the initial points of the random starting points).

localimprovements: An experimental technique that if it is true it will achieve very high quality results, at cost of increasing searching time.

vstate: A cache object for reducing memory allocations

Neighborhood algorithms

Regarding neighborhood algorithms, the package defines the following ones:

Neighborhood algorithms

SimilaritySearch.FixedNeighborhood — Type

FixedNeighborhood(k=8)

New items are connected with a fixed number of neighbors

SimilaritySearch.LogNeighborhood — Type

LogNeighborhood(b=2)

New items are connected with the $\log_b n$ items in the current graph (where $n$ is the current number of vertices in the graph)

SimilaritySearch.LogSatNeighborhood — Type

LogSatNeighborhood(b=1.1)

New items are connected with a half partitioning scheme based on the Spatial access tree (cite), $\log_b n$ near items are considered for this procedure. The Log-Sat method will produce small neighborhood with a nice covering of the search space.

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood(k=32)

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

Incremental construction of the index

The index supports insertions; for instance, it can be fully constructed using insertions as follows:

using SimilaritySearch

index = SearchGraph(L2Distance(), Vector{Float32}[])

for i in 1:1000
    push!(index, rand(Float32, 4))
end

search(index, rand(Float32, 4), 3)

KnnResult(3, 3, Item[Item(303, 0.15570286f0), Item(291, 0.17543355f0), Item(284, 0.19791986f0), Item(504, 0.17144701f0), Item(604, 0.17174023f0), Item(215, 0.1850284f0), Item(541, 0.18604517f0), Item(51, 0.1989268f0), Item(247, 0.20420702f0), Item(403, 0.21401183f0)])

Each insertion is performed by the push! function.

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

Customizing the insertion method

In particular, the procedure of pushing items into the index is made calling find_neighborhood and push_neighborhood; these methods can be overriden or used in other wyas to customize the insertion of data and the construction of the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(algo::FixedNeighborhood, index::SearchGraph, item)

Finds a list of neighbors using the FixedNeighborhood criterion of item in the index

find_neighborhood(index::SearchGraph{T}, item)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, L::AbstractVector{Int32})

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

Optimizing the index's performance

SimilaritySearch.optimize! — Function

function optimize!(search_algo::LocalSearchAlgorithm,
                   index::SearchGraph{T},
                   recall::Float64,
                   perf::Performance;
                   bsize::Int=4,
                   tol::Float64=0.01,
                   probes::Int=0) where T

Optimizes a local search index for an specific algorithm to get the desired performance. Note that optimizing for low-recall will yield to faster searches; the train queries are specified as part of the perf struct.