SearchGraph creates a graph

Tellez, E. S., Ruiz, G., Chavez, E., & Graff, M.A scalable solution to the nearest neighbor search problem through local-search methods on neighbor graphs. Pattern Analysis and Applications, 1-15.

As before, we need a dataset $X$ and a distance function.

using SimilaritySearch
X = [randn(3) for i in 1:10_000]
Q = [randn(3) for i in 1:5]

index = SearchGraph(; dist=L2Distance(), verbose=true)
append!(index, X)

for q in Q
    println(search(index, q, KnnResult(3)))
end

calling callback hints; n=10
calling callback neighborhood; n=10
calling callback hints; n=16
calling callback neighborhood; n=16
calling callback hints; n=24
calling callback neighborhood; n=24
calling callback hints; n=37
calling callback neighborhood; n=37
calling callback hints; n=56
calling callback neighborhood; n=56
calling callback hints; n=85
calling callback neighborhood; n=85
calling callback hints; n=128
calling callback neighborhood; n=128
calling callback hints; n=193
calling callback neighborhood; n=193
calling callback hints; n=290
calling callback neighborhood; n=290
calling callback hints; n=436
calling callback neighborhood; n=436
calling callback hints; n=655
calling callback neighborhood; n=655
calling callback hints; n=984
calling callback neighborhood; n=984
calling callback hints; n=1476
calling callback neighborhood; n=1476
calling callback hints; n=2215
calling callback neighborhood; n=2215
calling callback hints; n=3324
calling callback neighborhood; n=3324
calling callback hints; n=4986
calling callback neighborhood; n=4986
calling callback hints; n=7480
calling callback neighborhood; n=7480
KnnResult{Int32, Float32}(Int32[4195, 7138, 4318], Float32[0.059878666, 0.060887486, 0.06301622], 3, 0)
KnnResult{Int32, Float32}(Int32[8805, 7774, 501], Float32[0.05540494, 0.061832473, 0.07820696], 3, 0)
KnnResult{Int32, Float32}(Int32[3484, 2110, 8583], Float32[0.14745988, 0.15527269, 0.159456], 3, 0)
KnnResult{Int32, Float32}(Int32[8202, 7983, 3280], Float32[0.25871345, 0.43342942, 0.43885204], 3, 0)
KnnResult{Int32, Float32}(Int32[749, 8111, 7086], Float32[0.12272615, 0.14150539, 0.1450728], 3, 0)

There are a several options for the construction of the index, from specifying the precise searching algorithm (search_algo) and the neighborhood strategy to be used on construction (neighborhood_algo). You can also specify here if the construction will be made in parallel and how this paralllelism will be performed (firstblock, block).

SimilaritySearch.SearchGraph — Type

struct SearchGraph <: AbstractSearchContext

SearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.

Note 1: Parallel searches need copies of the structure or passing a KnnResult per thread. Note 2: Parallel insertions should be made through append! function with parallel=true

Missing docstring.

Missing docstring for SearchGraphOptions. Check Documenter's build log for details.

Local search algorithms

Regarding the search search algorithm, there are two alternatives, based on local search heuristics.

SimilaritySearch.BeamSearch — Type

BeamSearch(bsize::Integer=16, hints=Int32[], beam=KnnResult(bsize), vstate=VisitedVertices())

BeamSearch is an iteratively improving local search algorithm that explores the graph using blocks of bsize elements and neighborhoods at the time. Multithreading applications must have copies of this object due to shared cache objects.

hints: An initial hint for the exploration (empty hints imply bsize random starting points).

bsize: The size of the beam.

beam: A cache object for reducing memory allocations

vstate: A cache object for reducing memory allocations

SimilaritySearch.IHCSearch — Type

IHCSearch(hints::Vector; restarts=length(hints), localimprovements=false)
IHCSearch(restarts::Integer=20; hints=Int32[], localimprovements=false)
IHCSearch(hints, restarts, localimprovements, vstate)

IHCSearch is an iterated hill climbing algorithma, a local search algorithm. It greedily navigates the search graph and restart the search restarts times. Multithreading applications must have copies of this object due to shared cache objects.

restarts: The number of restarts.

hints: An initial hint for the exploration (if it is not empty, then superseeds the initial points of the random starting points).

localimprovements: An experimental technique that if it is true it will achieve very high quality results, at cost of increasing searching time.

vstate: A cache object for reducing memory allocations

Neighborhood algorithms

Regarding neighborhood algorithms, the package defines the following ones:

Neighborhood algorithms

Missing docstring.

Missing docstring for FixedNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogNeighborhood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LogSatNeighborhood. Check Documenter's build log for details.

SimilaritySearch.SatNeighborhood — Type

SatNeighborhood(k=32)

New items are connected with a small set of items computed with a SAT like scheme (cite). It starts with k near items that are reduced to a small neighborhood due to the SAT partitioning stage.

Incremental construction of the index

The index supports insertions; for instance, it can be fully constructed using insertions as follows:

using SimilaritySearch

index = SearchGraph()

append!(index, [rand(Float32, 4) for i in 1:1000])

search(index, rand(Float32, 4), 3)

KnnResult{Int32, Float32}(Int32[260, 681, 795], Float32[0.026340049, 0.027267773, 0.0303037], 3, 0)

Each insertion is performed by the push! function or a chunk of items with append!; the last function also allows parallel insertions.

Base.push! — Method

push!(index::SearchGraph, item)

Appends item into the index.

Base.append! — Method

append!(index::SearchGraph, db; parallel=false, parallel_firstblock=30_000, parallel_block=10_000, apply_callbacks=true)

Appends all items in db to the index. It can be made in parallel or sequentially. In case of a parallel appending, then parallel_firstblock indicates the minimum number of items before going parallel, and parallel_block sets the chunck size to append in parallel.

Note: Parallel construction doesn't trigger callbacks listed in `callbacks', they must be executed manually.

Customizing the insertion method

In particular, the procedure of pushing items into the index is made calling find_neighborhood and push_neighborhood; these methods can be overriden or used in other wyas to customize the insertion of data and the construction of the index.

SimilaritySearch.find_neighborhood — Function

find_neighborhood(index::SearchGraph{T}, item; res=index.res)

Searches for item neighborhood in the index, i.e., if item were in the index whose items should be its neighbors (intenal function)

SimilaritySearch.push_neighborhood! — Function

push_neighborhood!(index::SearchGraph, item, neighbors::KnnResult; apply_callbacks=true)

Inserts the object item into the index, i.e., creates an edge from items listed in L and the vertex created for ìtem` (internal function)

Optimizing the index's performance

Missing docstring.

Missing docstring for optimize!. Check Documenter's build log for details.